wgpu::api

Module bind_group

Structs

BindGroup

Handle to a binding group.

BindGroupDescriptor

Describes a group of bindings and the resources to be bound.

BindGroupEntry

An element of a BindGroupDescriptor, consisting of a bindable resource and the slot to bind it to.

BufferBinding

Describes the segment of a buffer to bind.

Enums

BindingResource

Resource to be bound by a BindGroup for use with a pipeline.