wgpu::api
Module blas

Structs

Blas

Bottom Level Acceleration Structure (BLAS).

BlasBuildEntry

Builds the given sets of geometry into the given Blas.

BlasTriangleGeometry

Definition for a triangle geometry for a Bottom Level Acceleration Structure (BLAS).

ContextBlasBuildEntry

Context version see BlasBuildEntry.

ContextBlasTriangleGeometry

Context version of BlasTriangleGeometry.

TlasInstance

Safe instance for a Tlas.

Enums

BlasGeometries

Contains the sets of geometry that go into a Blas.

ContextBlasGeometries

Context version of BlasGeometries.

Type Aliases

AccelerationStructureFlags

Flags for an acceleration structure.

AccelerationStructureGeometryFlags

Flags for a geometry inside a bottom level acceleration structure.

AccelerationStructureUpdateMode

Update mode for acceleration structure builds.

BlasGeometrySizeDescriptors

Descriptor for the size defining attributes, for a bottom level acceleration structure.

BlasTriangleGeometrySizeDescriptor

Descriptor for the size defining attributes of a triangle geometry, for a bottom level acceleration structure.

CreateBlasDescriptor

Descriptor to create bottom level acceleration structures.