wgpu::api::buffer

Struct BufferViewMut

pub struct BufferViewMut {
    buffer: Buffer,
    offset: BufferAddress,
    size: BufferSize,
    inner: DispatchBufferMappedRange,
}

A write-only view of a mapped buffer’s bytes.

To get a BufferViewMut, first map the buffer, and then call buffer.slice(range).get_mapped_range_mut().

Because Rust has no write-only reference type (&[u8] is read-only and &mut [u8] is read-write), this type does not dereference to a slice in the way that BufferView does. Instead, .slice() returns a special WriteOnly pointer type, and there are also a few convenience methods such as BufferViewMut::copy_from_slice().

Before the buffer can be unmapped, all BufferViewMuts observing it must be dropped. Otherwise, the call to Buffer::unmap will panic.

For example code, see the documentation on mapping buffers.

Fields
buffer: Bufferoffset: BufferAddresssize: BufferSizeinner: DispatchBufferMappedRange
Implementations

impl BufferViewMut

These methods are equivalent to the methods of the same names on WriteOnly.

pub fn len(&self) -> usize

Returns the length of this view; the number of bytes to be written.

pub fn is_empty(&self) -> bool

Returns true if the view has a length of 0.

Note that this is currently impossible.

pub fn slice<'a, S: RangeBounds<usize>>( &'a mut self, bounds: S, ) -> WriteOnly<'a, [u8]>

Returns a WriteOnly reference to a portion of this.

.slice(..) can be used to access the whole data.

pub fn copy_from_slice(&mut self, src: &[u8])

Copies all elements from src into self.

The length of src must be the same as self.

This method is equivalent to self.slice(..).copy_from_slice(src).