wgpu::api

Module command_buffer_actions

Not a root type, but common types for command buffer deferral actions.

Macros

impl_deferred_command_buffer_actions 🔒

Structs

DeferredBufferMapping 🔒

A deferred buffer mapping request captured during encoding (or a pass) and executed later when the command buffer is submitted.

DeferredCommandBufferActions 🔒

Set of actions to take when the command buffer is submitted.

Type Aliases

SharedDeferredCommandBufferActions 🔒