wgpu::api
Module command_encoder

Structs

CommandEncoder

Encodes a series of GPU operations.

TexelCopyBufferInfoBase

View of a buffer which can be used to copy to/from a texture.

TexelCopyTextureInfoBase

View of a texture which can be used to copy to/from a buffer/texture.

Type Aliases

CommandEncoderDescriptor

Describes a CommandEncoder.

TexelCopyBufferInfo

View of a buffer which can be used to copy to/from a texture.

TexelCopyTextureInfo

View of a texture which can be used to copy to/from a buffer/texture.