wgpu::api::device

Type Alias DeviceDescriptor

pub type DeviceDescriptor<'a> = DeviceDescriptor<Label<'a>>;

Describes a Device.

For use with Adapter::request_device.

Corresponds to WebGPU GPUDeviceDescriptor.

Aliased Type

struct DeviceDescriptor<'a> {
    pub label: Option<&'a str>,
    pub required_features: Features,
    pub required_limits: Limits,
    pub experimental_features: ExperimentalFeatures,
    pub memory_hints: MemoryHints,
    pub trace: Trace,
}

Fields
label: Option<&'a str>

Debug label for the device.

required_features: Features

Specifies the features that are required by the device request. The request will fail if the adapter cannot provide these features.

Exactly the specified set of features, and no more or less, will be allowed in validation of API calls on the resulting device.

required_limits: Limits

Specifies the limits that are required by the device request. The request will fail if the adapter cannot provide these limits.

Exactly the specified limits, and no better or worse, will be allowed in validation of API calls on the resulting device.

experimental_features: ExperimentalFeatures

Specifies whether self.required_features is allowed to contain experimental features.

memory_hints: MemoryHints

Hints for memory allocation strategies.

trace: Trace

Whether API tracing for debugging is enabled, and where the trace is written if so.