wgpu::api
Module queue

Structs

Queue

Handle to a command queue on a device.

QueueWriteBufferView

A write-only view into a staging buffer.

SubmissionIndex

Identifier for a particular call to Queue::submit. Can be used as part of an argument to Device::poll to block for a particular submission to finish.

Enums

MaintainBase

Passed to Device::poll to control how and if it should block.

Type Aliases

PollType

Passed to Device::poll to control how and if it should block.