wgpu::api::render_pass

Struct RenderPass

pub struct RenderPass<'encoder> {
    pub(crate) inner: DispatchRenderPass,
    pub(crate) actions: Arc<Mutex<DeferredCommandBufferActions>>,
    pub(crate) _encoder_guard: PhantomDrop<&'encoder ()>,
}

In-progress recording of a render pass: a list of render commands in a CommandEncoder.

It can be created with CommandEncoder::begin_render_pass(), whose RenderPassDescriptor specifies the attachments (textures) that will be rendered to.

Most of the methods on RenderPass serve one of two purposes, identifiable by their names:

draw_*(): Drawing (that is, encoding a render command, which, when executed by the GPU, will rasterize something and execute shaders).

set_*(): Setting part of the render state for future drawing commands.

A render pass may contain any number of drawing commands, and before/between each command the render state may be updated however you wish; each drawing command will be executed using the render state that has been set when the draw_*() function is called.

Corresponds to WebGPU GPURenderPassEncoder.

Fields
inner: DispatchRenderPassactions: Arc<Mutex<DeferredCommandBufferActions>>_encoder_guard: PhantomDrop<&'encoder ()>

This lifetime is used to protect the CommandEncoder from being used while the pass is alive. This needs to be PhantomDrop to prevent the lifetime from being shortened.

Implementations

impl RenderPass<'_>

pub fn forget_lifetime(self) -> RenderPass<'static>

Drops the lifetime relationship to the parent command encoder, making usage of the encoder while this pass is recorded a run-time error instead.

Attention: As long as the render pass has not been ended, any mutating operation on the parent command encoder will cause a run-time error and invalidate it! By default, the lifetime constraint prevents this, but it can be useful to handle this at run time, such as when storing the pass and encoder in the same data structure.

This operation has no effect on pass recording. It’s a safe operation, since CommandEncoder is in a locked state as long as the pass is active regardless of the lifetime constraint or its absence.

pub fn set_bind_group<'a, BG>( &mut self, index: u32, bind_group: BG, offsets: &[DynamicOffset], )
where Option<&'a BindGroup>: From<BG>,

Sets the active bind group for a given bind group index. The bind group layout in the active pipeline when any draw_*() method is called must match the layout of this bind group.

If the bind group have dynamic offsets, provide them in binding order. These offsets have to be aligned to Limits::min_uniform_buffer_offset_alignment or Limits::min_storage_buffer_offset_alignment appropriately.

Subsequent draw calls’ shader executions will be able to access data in these bind groups.

pub fn set_pipeline(&mut self, pipeline: &RenderPipeline)

Sets the active render pipeline.

Subsequent draw calls will exhibit the behavior defined by pipeline.

pub fn set_blend_constant(&mut self, color: Color)

Sets the blend color as used by some of the blending modes.

Subsequent blending tests will test against this value. If this method has not been called, the blend constant defaults to Color::TRANSPARENT (all components zero).

pub fn set_index_buffer( &mut self, buffer_slice: BufferSlice<'_>, index_format: IndexFormat, )

Sets the active index buffer.

Subsequent calls to draw_indexed on this RenderPass will use buffer as the source index buffer.

pub fn set_vertex_buffer(&mut self, slot: u32, buffer_slice: BufferSlice<'_>)

Assign a vertex buffer to a slot.

Subsequent calls to draw and draw_indexed on this RenderPass will use buffer as one of the source vertex buffers. The format of the data in the buffer is specified by the VertexBufferLayout in the pipeline’s VertexState.

The slot refers to the index of the matching descriptor in VertexState::buffers.

pub fn set_scissor_rect(&mut self, x: u32, y: u32, width: u32, height: u32)

Sets the scissor rectangle used during the rasterization stage. After transformation into viewport coordinates.

Subsequent draw calls will discard any fragments which fall outside the scissor rectangle. If this method has not been called, the scissor rectangle defaults to the entire bounds of the render targets.

The function of the scissor rectangle resembles set_viewport(), but it does not affect the coordinate system, only which fragments are discarded.

pub fn set_viewport( &mut self, x: f32, y: f32, w: f32, h: f32, min_depth: f32, max_depth: f32, )

Sets the viewport used during the rasterization stage to linearly map from normalized device coordinates to viewport coordinates.

Subsequent draw calls will only draw within this region. If this method has not been called, the viewport defaults to the entire bounds of the render targets.

pub fn set_stencil_reference(&mut self, reference: u32)

Sets the stencil reference.

Subsequent stencil tests will test against this value. If this method has not been called, the stencil reference value defaults to 0.

pub fn insert_debug_marker(&mut self, label: &str)

Inserts debug marker.

pub fn push_debug_group(&mut self, label: &str)

Start record commands and group it into debug marker group.

pub fn pop_debug_group(&mut self)

Stops command recording and creates debug group.

pub fn draw(&mut self, vertices: Range<u32>, instances: Range<u32>)

Draws primitives from the active vertex buffer(s).

The active vertex buffer(s) can be set with RenderPass::set_vertex_buffer. This does not use an index buffer. If you need indexed drawing, see RenderPass::draw_indexed

Panics if vertices range is outside of the range of the vertices range of any set vertex buffer.

vertices: The range of vertices to draw.

instances: Range of instances to draw. Use 0..1 if instance buffers are not used.

E.g.of how its used internally

for instance_id in instance_range {
    for vertex_id in vertex_range {
        let vertex = vertex[vertex_id];
        vertex_shader(vertex, vertex_id, instance_id);
    }
}

This drawing command uses the current render state, as set by preceding set_*() methods. It is not affected by changes to the state that are performed after it is called.

pub fn draw_indexed( &mut self, indices: Range<u32>, base_vertex: i32, instances: Range<u32>, )

Draws indexed primitives using the active index buffer and the active vertex buffers.

The active index buffer can be set with RenderPass::set_index_buffer The active vertex buffers can be set with RenderPass::set_vertex_buffer.

Panics if indices range is outside of the range of the indices range of the set index buffer.

indices: The range of indices to draw.

base_vertex: value added to each index value before indexing into the vertex buffers.

instances: Range of instances to draw. Use 0..1 if instance buffers are not used.

E.g.of how its used internally

for instance_id in instance_range {
    for index_index in index_range {
        let vertex_id = index_buffer[index_index];
        let adjusted_vertex_id = vertex_id + base_vertex;
        let vertex = vertex[adjusted_vertex_id];
        vertex_shader(vertex, adjusted_vertex_id, instance_id);
    }
}

This drawing command uses the current render state, as set by preceding set_*() methods. It is not affected by changes to the state that are performed after it is called.

pub fn draw_mesh_tasks( &mut self, group_count_x: u32, group_count_y: u32, group_count_z: u32, )

Draws using a mesh pipeline.

The current pipeline must be a mesh pipeline.

If the current pipeline has a task shader, run it with an workgroup for every vec3<u32>(i, j, k) where i, j, and k are between 0 and group_count_x, group_count_y, and group_count_z. The invocation with index zero in each group is responsible for determining the mesh shader dispatch. Its return value indicates the number of workgroups of mesh shaders to invoke. It also passes a payload value for them to consume. Because each task workgroup is essentially a mesh shader draw call, mesh workgroups dispatched by different task workgroups cannot interact in any way, and workgroup_id corresponds to its location in the calling specific task shader’s dispatch group.

If the current pipeline lacks a task shader, run its mesh shader with a workgroup for every vec3<u32>(i, j, k) where i, j, and k are between 0 and group_count_x, group_count_y, and group_count_z.

Each mesh shader workgroup outputs a set of vertices and indices for primitives. The indices outputted correspond to the vertices outputted by that same workgroup; there is no global vertex buffer. These primitives are passed to the rasterizer and essentially treated like a vertex shader output, except that the mesh shader may choose to cull specific primitives or pass per-primitive non-interpolated values to the fragment shader. As such, each primitive is then rendered with the current pipeline’s fragment shader, if present. Otherwise, No Color Output mode is used.

pub fn draw_indirect( &mut self, indirect_buffer: &Buffer, indirect_offset: BufferAddress, )

Draws primitives from the active vertex buffer(s) based on the contents of the indirect_buffer.

This is like calling RenderPass::draw but the contents of the call are specified in the indirect_buffer. The structure expected in indirect_buffer must conform to DrawIndirectArgs.

Calling this requires the device support DownlevelFlags::INDIRECT_EXECUTION.

pub fn draw_indexed_indirect( &mut self, indirect_buffer: &Buffer, indirect_offset: BufferAddress, )

Draws indexed primitives using the active index buffer and the active vertex buffers, based on the contents of the indirect_buffer.

This is like calling RenderPass::draw_indexed but the contents of the call are specified in the indirect_buffer. The structure expected in indirect_buffer must conform to DrawIndexedIndirectArgs.

Calling this requires the device support DownlevelFlags::INDIRECT_EXECUTION.

pub fn draw_mesh_tasks_indirect( &mut self, indirect_buffer: &Buffer, indirect_offset: BufferAddress, )

Draws using a mesh pipeline, based on the contents of the indirect_buffer

This is like calling RenderPass::draw_mesh_tasks but the contents of the call are specified in the indirect_buffer. The structure expected in the indirect_buffer must conform to DispatchIndirectArgs.

Indirect drawing has some caveats depending on the features available. We are not currently able to validate these and issue an error.

See details on the individual flags for more information.

pub fn map_buffer_on_submit<S: RangeBounds<BufferAddress>>( &self, buffer: &Buffer, mode: MapMode, bounds: S, callback: impl FnOnce(Result<(), BufferAsyncError>) + WasmNotSend + 'static, )

On submission, maps the buffer to host (CPU) memory, making it available for reading or writing via get_mapped_range(). The buffer becomes accessible once the callback is invoked with Ok.

Use this when you need to submit work that uses the buffer before mapping it. Because that submission must happen before calling map_async, this method schedules the mapping for after submission, avoiding extra calls to Buffer::map_async() or BufferSlice::map_async() and letting you start the mapping from a more convenient place.

For the callback to run, either queue.submit(..), instance.poll_all(..), or device.poll(..) must be called elsewhere in the runtime, possibly integrated into an event loop or run on a separate thread.

The callback runs on the thread that first calls one of the above functions after the GPU work completes. There are no restrictions on the code you can run in the callback; however, on native the polling call will not return until the callback finishes, so keep callbacks short (set flags, send messages, etc.).

While a buffer is mapped, it cannot be used by other commands; at any time, either the GPU or the CPU has exclusive access to the buffer’s contents.

Panics

If bounds is outside the bounds of buffer.

If bounds has a length less than 1.

Panics During Submit

If the buffer is already mapped.

If the buffer’s BufferUsages do not allow the requested MapMode.

If bounds is outside of the bounds of buffer.

If bounds does not start at a multiple of MAP_ALIGNMENT.

If bounds has a length that is not a multiple of 4 greater than 0.

pub fn on_submitted_work_done(&self, callback: impl FnOnce() + Send + 'static)

Registers a callback that is invoked when this command buffer’s work finishes executing on the GPU. When this callback runs, all mapped-buffer callbacks registered for the same submission are guaranteed to have been called.

For the callback to run, either queue.submit(..), instance.poll_all(..), or device.poll(..) must be called elsewhere in the runtime, possibly integrated into an event loop or run on a separate thread.

The callback runs on the thread that first calls one of the above functions after the GPU work completes. There are no restrictions on the code you can run in the callback; however, on native the polling call will not return until the callback finishes, so keep callbacks short (set flags, send messages, etc.).

pub fn execute_bundles<'a, I: IntoIterator<Item = &'a RenderBundle>>( &mut self, render_bundles: I, )

Execute a render bundle, which is a set of pre-recorded commands that can be run together.

Commands in the bundle do not inherit this render pass’s current render state, and after the bundle has executed, the state is cleared (reset to defaults, not the previous state).

pub fn multi_draw_indirect( &mut self, indirect_buffer: &Buffer, indirect_offset: BufferAddress, count: u32, )

Dispatches multiple draw calls from the active vertex buffer(s) based on the contents of the indirect_buffer. count draw calls are issued.

The active vertex buffers can be set with RenderPass::set_vertex_buffer.

The structure expected in indirect_buffer must conform to DrawIndirectArgs. These draw structures are expected to be tightly packed.

Calling this requires the device support DownlevelFlags::INDIRECT_EXECUTION.

This drawing command uses the current render state, as set by preceding set_*() methods. It is not affected by changes to the state that are performed after it is called.

pub fn multi_draw_indexed_indirect( &mut self, indirect_buffer: &Buffer, indirect_offset: BufferAddress, count: u32, )

Dispatches multiple draw calls from the active index buffer and the active vertex buffers, based on the contents of the indirect_buffer. count draw calls are issued.

The active index buffer can be set with RenderPass::set_index_buffer, while the active vertex buffers can be set with RenderPass::set_vertex_buffer.

The structure expected in indirect_buffer must conform to DrawIndexedIndirectArgs. These draw structures are expected to be tightly packed.

Calling this requires the device support DownlevelFlags::INDIRECT_EXECUTION.

This drawing command uses the current render state, as set by preceding set_*() methods. It is not affected by changes to the state that are performed after it is called.

pub fn multi_draw_mesh_tasks_indirect( &mut self, indirect_buffer: &Buffer, indirect_offset: BufferAddress, count: u32, )

Dispatches multiple draw calls based on the contents of the indirect_buffer. count draw calls are issued.

The structure expected in the indirect_buffer must conform to DispatchIndirectArgs.

This drawing command uses the current render state, as set by preceding set_*() methods. It is not affected by changes to the state that are performed after it is called.

pub fn as_custom<T: RenderPassInterface>(&self) -> Option<&T>

Returns custom implementation of RenderPass (if custom backend and is internally T)

impl RenderPass<'_>

Features::MULTI_DRAW_INDIRECT_COUNT must be enabled on the device in order to call these functions.

pub fn multi_draw_indirect_count( &mut self, indirect_buffer: &Buffer, indirect_offset: BufferAddress, count_buffer: &Buffer, count_offset: BufferAddress, max_count: u32, )

Dispatches multiple draw calls from the active vertex buffer(s) based on the contents of the indirect_buffer. The count buffer is read to determine how many draws to issue.

The indirect buffer must be long enough to account for max_count draws, however only count draws will be read. If count is greater than max_count, max_count will be used.

The active vertex buffers can be set with RenderPass::set_vertex_buffer.

The structure expected in indirect_buffer must conform to DrawIndirectArgs. These draw structures are expected to be tightly packed.

The structure expected in count_buffer is the following:

#[repr(C)]
struct DrawIndirectCount {
    count: u32, // Number of draw calls to issue.
}

This drawing command uses the current render state, as set by preceding set_*() methods. It is not affected by changes to the state that are performed after it is called.

pub fn multi_draw_indexed_indirect_count( &mut self, indirect_buffer: &Buffer, indirect_offset: BufferAddress, count_buffer: &Buffer, count_offset: BufferAddress, max_count: u32, )

Dispatches multiple draw calls from the active index buffer and the active vertex buffers, based on the contents of the indirect_buffer. The count buffer is read to determine how many draws to issue.

The indirect buffer must be long enough to account for max_count draws, however only count draws will be read. If count is greater than max_count, max_count will be used.

The active index buffer can be set with RenderPass::set_index_buffer, while the active vertex buffers can be set with RenderPass::set_vertex_buffer.

The structure expected in indirect_buffer must conform to DrawIndexedIndirectArgs.

These draw structures are expected to be tightly packed.

The structure expected in count_buffer is the following:

#[repr(C)]
struct DrawIndexedIndirectCount {
    count: u32, // Number of draw calls to issue.
}

This drawing command uses the current render state, as set by preceding set_*() methods. It is not affected by changes to the state that are performed after it is called.

pub fn multi_draw_mesh_tasks_indirect_count( &mut self, indirect_buffer: &Buffer, indirect_offset: BufferAddress, count_buffer: &Buffer, count_offset: BufferAddress, max_count: u32, )

Dispatches multiple draw calls based on the contents of the indirect_buffer. The count buffer is read to determine how many draws to issue.

The indirect buffer must be long enough to account for max_count draws, however only count draws will be read. If count is greater than max_count, max_count will be used.

The structure expected in the indirect_buffer must conform to DispatchIndirectArgs.

These draw structures are expected to be tightly packed.

This drawing command uses the current render state, as set by preceding set_*() methods. It is not affected by changes to the state that are performed after it is called.

impl RenderPass<'_>

Features::IMMEDIATES must be enabled on the device in order to call these functions.

pub fn set_immediates(&mut self, stages: ShaderStages, offset: u32, data: &[u8])

Set immediate data data for subsequent draw calls.

Write the bytes in data at offset offset within immediate data storage, all of which are accessible by all the pipeline stages in stages, and no others. Both offset and the length of data must be multiples of IMMEDIATES_ALIGNMENT, which is always 4.

For example, if offset is 4 and data is eight bytes long, this call will write data to bytes 4..12 of immediate data storage.

Stage matching

Every byte in the affected range of immediate data storage must be accessible to exactly the same set of pipeline stages, which must match stages. If there are two bytes of storage that are accessible by different sets of pipeline stages - say, one is accessible by fragment shaders, and the other is accessible by both fragment shaders and vertex shaders - then no single set_immediates call may affect both of them; to write both, you must make multiple calls, each with the appropriate stages value.

Which pipeline stages may access a given byte is determined by the pipeline’s Immediate global variable and (if it is a struct) its members’ offsets.

For example, suppose you have twelve bytes of immediate data storage, where bytes 0..8 are accessed by the vertex shader, and bytes 4..12 are accessed by the fragment shader. This means there are three byte ranges each accessed by a different set of stages:

Bytes 0..4 are accessed only by the fragment shader.

Bytes 4..8 are accessed by both the fragment shader and the vertex shader.

Bytes 8..12 are accessed only by the vertex shader.

To write all twelve bytes requires three set_immediates calls, one for each range, each passing the matching stages mask.

impl RenderPass<'_>

Features::TIMESTAMP_QUERY_INSIDE_PASSES must be enabled on the device in order to call these functions.

pub fn write_timestamp(&mut self, query_set: &QuerySet, query_index: u32)

Issue a timestamp command at this point in the queue. The timestamp will be written to the specified query set, at the specified index.

Must be multiplied by Queue::get_timestamp_period to get the value in nanoseconds. Absolute values have no meaning, but timestamps can be subtracted to get the time it takes for a string of operations to complete.

impl RenderPass<'_>

pub fn begin_occlusion_query(&mut self, query_index: u32)

Start a occlusion query on this render pass. It can be ended with end_occlusion_query. Occlusion queries may not be nested.

pub fn end_occlusion_query(&mut self)

End the occlusion query on this render pass. It can be started with begin_occlusion_query. Occlusion queries may not be nested.

impl RenderPass<'_>

Features::PIPELINE_STATISTICS_QUERY must be enabled on the device in order to call these functions.

pub fn begin_pipeline_statistics_query( &mut self, query_set: &QuerySet, query_index: u32, )

Start a pipeline statistics query on this render pass. It can be ended with end_pipeline_statistics_query. Pipeline statistics queries may not be nested.

pub fn end_pipeline_statistics_query(&mut self)

End the pipeline statistics query on this render pass. It can be started with begin_pipeline_statistics_query. Pipeline statistics queries may not be nested.