wgpu::api

Module render_pipeline

Structs

FragmentState

Describes the fragment processing in a render pipeline.

MeshPipelineDescriptor

Describes a mesh shader (graphics) pipeline.

MeshState

Describes the mesh shader stage in a mesh shader pipeline.

RenderPipeline

Handle to a rendering (graphics) pipeline.

RenderPipelineDescriptor

Describes a render (graphics) pipeline.

TaskState

Describes the task shader stage in a mesh shader pipeline.

VertexBufferLayout

Specifies an interpretation of the bytes of a vertex buffer as vertex attributes.

VertexState

Describes the vertex processing in a render pipeline.