wgpu::api

Module shader_module

Structs

CompilationInfo

Compilation information for a shader module.

CompilationMessage

A single message from the shader compilation process.

ShaderModule

Handle to a compiled shader module.

ShaderModuleDescriptor

Descriptor for use with Device::create_shader_module.

SourceLocation

A human-readable representation for a span, tailored for text source.

Enums

CompilationMessageType

The type of a compilation message.

ShaderSource

Source of a shader module.

Type Aliases

ShaderModuleDescriptorDxil

Descriptor for a shader module given by DirectX DXIL source.

ShaderModuleDescriptorHlsl

Descriptor for a shader module given by DirectX HLSL source.

ShaderModuleDescriptorMsl

Descriptor for a shader module given by Metal MSL source.

ShaderModuleDescriptorPassthrough

Descriptor for a shader module that will bypass wgpu’s shader tooling, for use with Device::create_shader_module_passthrough.

ShaderModuleDescriptorSpirV

Descriptor for a shader module given by SPIR-V binary.