wgpu::api
Module surface

Structs

CreateSurfaceError

Instance::create_surface() or a related function failed.

Surface

Handle to a presentable surface.

Enums

CreateSurfaceErrorKind ๐

SurfaceTarget

The window/canvas/surface/swap-chain/etc. a surface is attached to, for use with safe surface creation.

SurfaceTargetUnsafe

The window/canvas/surface/swap-chain/etc. a surface is attached to, for use with unsafe surface creation.

Traits

WindowHandle

Super trait for window handles as used in SurfaceTarget.

Type Aliases

SurfaceConfiguration

Describes a Surface.