Struct wgpu::core::hub::Hub

pub struct Hub<A>
where
    A: HalApi,
{

    pub adapters: Registry<Adapter<A>>,
    pub devices: Registry<Device<A>>,
    pub queues: Registry<Queue<A>>,
    pub pipeline_layouts: Registry<PipelineLayout<A>>,
    pub shader_modules: Registry<ShaderModule<A>>,
    pub bind_group_layouts: Registry<BindGroupLayout<A>>,
    pub bind_groups: Registry<BindGroup<A>>,
    pub command_buffers: Registry<CommandBuffer<A>>,
    pub render_bundles: Registry<RenderBundle<A>>,
    pub render_pipelines: Registry<RenderPipeline<A>>,
    pub compute_pipelines: Registry<ComputePipeline<A>>,
    pub query_sets: Registry<QuerySet<A>>,
    pub buffers: Registry<Buffer<A>>,
    pub staging_buffers: Registry<StagingBuffer<A>>,
    pub textures: Registry<Texture<A>>,
    pub texture_views: Registry<TextureView<A>>,
    pub samplers: Registry<Sampler<A>>,

}

Available on wgpu_core only.

All the resources for a particular backend in a crate::global::Global.

To obtain global’s Hub for some HalApi backend type A, call A::hub(global).

Locking

Each field in Hub is a Registry holding all the values of a particular type of resource, all protected by a single RwLock. So for example, to access any Buffer, you must acquire a read lock on the Hubs entire buffers registry. The lock guard gives you access to the Registry’s Storage, which you can then index with the buffer’s id. (Yes, this design causes contention; see [#2272].)

But most wgpu operations require access to several different kinds of resource, so you often need to hold locks on several different fields of your Hub simultaneously.

Inside the Registry there are Arc<T> where T is a Resource Lock of Registry happens only when accessing to get the specific resource

Fields
adapters: Registry<Adapter<A>>devices: Registry<Device<A>>queues: Registry<Queue<A>>pipeline_layouts: Registry<PipelineLayout<A>>shader_modules: Registry<ShaderModule<A>>bind_group_layouts: Registry<BindGroupLayout<A>>bind_groups: Registry<BindGroup<A>>command_buffers: Registry<CommandBuffer<A>>render_bundles: Registry<RenderBundle<A>>render_pipelines: Registry<RenderPipeline<A>>compute_pipelines: Registry<ComputePipeline<A>>query_sets: Registry<QuerySet<A>>buffers: Registry<Buffer<A>>staging_buffers: Registry<StagingBuffer<A>>textures: Registry<Texture<A>>texture_views: Registry<TextureView<A>>samplers: Registry<Sampler<A>>
Implementations

impl<A> Hub<A>
where A: HalApi,

pub fn generate_report(&self) -> HubReport