wgpu::core

Module validation

Structs

Interface

InterfaceVar

MissingBufferUsageError

MissingTextureUsageError

NumericType

Enums

BindingError

BindingLayoutSource

FilteringError

InputError

StageError

Errors produced when validating a programmable stage of a pipeline.

Functions

check_buffer_usage

Checks that the given buffer usage contains the required buffer usage, returns an error otherwise.

check_texture_format

Return true if the fragment format is covered by the provided output.

check_texture_usage

Checks that the given texture usage contains the required texture usage, returns an error otherwise.

Type Aliases

StageIo