wgpu::documentation::internals

Module reporting_driver_bugs

Reporting Driver Bugs

It can sometimes be hard to figure out where to report driver bugs to the vendor. Here is a collection of known places for issues to go, and whether reports have been successful.

Proprietary:

NVIDIA Windows/Linux: forum (success)

AMD (Windows — GL/VK): forum (unknown)

Intel (Windows): GitHub (unknown)

Qualcomm: forum (success)

WARP: ping @soldieroflight in #dx12-developers on the DX12 Discord

Open source:

AMDVLK (Linux): GitHub

Mesa: GitLab

AMD (Linux) RADV

Intel (Linux)

Raspberry Pi

llvmpipe

lavapipe

MoltenVK: GitHub