Enum wgpu::BindingType

pub enum BindingType {
    UniformBuffer {
        dynamic: bool,
        min_binding_size: Option<NonZeroU64>,
    },
    StorageBuffer {
        dynamic: bool,
        min_binding_size: Option<NonZeroU64>,
        readonly: bool,
    },
    Sampler {
        comparison: bool,
    },
    SampledTexture {
        dimension: TextureViewDimension,
        component_type: TextureComponentType,
        multisampled: bool,
    },
    StorageTexture {
        dimension: TextureViewDimension,
        format: TextureFormat,
        readonly: bool,
    },
}

Specific type of a binding.

WebGPU spec: https://gpuweb.github.io/gpuweb/#dictdef-gpubindgrouplayoutentry

Variants

UniformBuffer

A buffer for uniform values.

Example GLSL syntax:

layout(std140, binding = 0)
uniform Globals {
    vec2 aUniform;
    vec2 anotherUniform;
};

Fields of UniformBuffer

dynamic: bool

Indicates that the binding has a dynamic offset. One offset must be passed to [RenderPass::set_bind_group] for each dynamic binding in increasing order of binding number.

min_binding_size: Option<NonZeroU64>

Minimum size of the corresponding BufferBinding required to match this entry. When pipeline is created, the size has to cover at least the corresponding structure in the shader plus one element of the unbound array, which can only be last in the structure. If None, the check is performed at draw call time instead of pipeline and bind group creation.

StorageBuffer

A storage buffer.

Example GLSL syntax:

layout (set=0, binding=0) buffer myStorageBuffer {
    vec4 myElement[];
};

Fields of StorageBuffer

dynamic: bool

Indicates that the binding has a dynamic offset. One offset must be passed to [RenderPass::set_bind_group] for each dynamic binding in increasing order of binding number.

min_binding_size: Option<NonZeroU64>

Minimum size of the corresponding BufferBinding required to match this entry. When pipeline is created, the size has to cover at least the corresponding structure in the shader plus one element of the unbound array, which can only be last in the structure. If None, the check is performed at draw call time instead of pipeline and bind group creation.

readonly: bool

The buffer can only be read in the shader and it must be annotated with readonly.

Example GLSL syntax:

layout (set=0, binding=0) readonly buffer myStorageBuffer {
    vec4 myElement[];
};

Sampler

A sampler that can be used to sample a texture.

Example GLSL syntax:

layout(binding = 0)
uniform sampler s;

Fields of Sampler

comparison: bool

Use as a comparison sampler instead of a normal sampler. For more info take a look at the analogous functionality in OpenGL: https://www.khronos.org/opengl/wiki/Sampler_Object#Comparison_mode.

SampledTexture

A texture.

Example GLSL syntax:

layout(binding = 0)
uniform texture2D t;

Fields of SampledTexture

dimension: TextureViewDimension

Dimension of the texture view that is going to be sampled.

component_type: TextureComponentType

Component type of the texture. This must be compatible with the format of the texture.

multisampled: bool

True if the texture has a sample count greater than 1. If this is true, the texture must be read from shaders with texture1DMS, texture2DMS, or texture3DMS, depending on dimension.

StorageTexture

A storage texture.

Example GLSL syntax:

layout(set=0, binding=0, r32f) uniform image2D myStorageImage;

Note that the texture format must be specified in the shader as well. A list of valid formats can be found in the specification here: https://www.khronos.org/registry/OpenGL/specs/gl/GLSLangSpec.4.60.html#layout-qualifiers

Fields of StorageTexture

dimension: TextureViewDimension

Dimension of the texture view that is going to be sampled.

format: TextureFormat

Format of the texture.

readonly: bool

The texture can only be read in the shader and it must be annotated with readonly.

Example GLSL syntax:

layout(set=0, binding=0, r32f) readonly uniform image2D myStorageImage;

Implementations

impl BindingType

pub fn has_dynamic_offset(&self) -> bool