wgpu
Enum Dx12Compiler

pub enum Dx12Compiler {
    Fxc,
    DynamicDxc {
        dxc_path: String,
        dxil_path: String,
        max_shader_model: DxcShaderModel,
    },
    StaticDxc,
}

Selects which DX12 shader compiler to use.

If the DynamicDxc option is selected, but dxcompiler.dll and dxil.dll files arenโt found, then this will fall back to the Fxc compiler at runtime and log an error.

Variants

Fxc

The Fxc compiler (default) is old, slow and unmaintained.

However, it doesnโt require any additional .dlls to be shipped with the application.

DynamicDxc

The Dxc compiler is new, fast and maintained.

However, it requires both dxcompiler.dll and dxil.dll to be shipped with the application. These files can be downloaded from https://github.com/microsoft/DirectXShaderCompiler/releases.

Minimum supported version: v1.5.2010

It also requires WDDM 2.1 (Windows 10 version 1607).

Fields

dxc_path: String

Path to dxcompiler.dll.

dxil_path: String

Path to dxil.dll.

max_shader_model: DxcShaderModel

Maximum shader model the given dll supports.

StaticDxc

The statically-linked variant of Dxc.

The static-dxc feature is required for this setting to be used successfully on DX12. Not available on windows-aarch64-pc-* targets.

Implementations

impl Dx12Compiler

pub fn default_dynamic_dxc() -> Dx12Compiler

Helper function to construct a DynamicDxc variant with default paths.

The dll must support at least shader model 6.5.

pub fn from_env() -> Option<Dx12Compiler>

Choose which DX12 shader compiler to use from the environment variable WGPU_DX12_COMPILER.

Valid values, case insensitive:

Fxc

Dxc or DynamicDxc

StaticDxc

pub fn with_env(self) -> Dx12Compiler

Takes the given compiler, modifies it based on the WGPU_DX12_COMPILER environment variable, and returns the result.

See from_env for more information.