Trait wgpu::hal::Adapter

pub trait Adapter<A>: WasmNotSendSync
where
    A: Api,
{
    // Required methods
    unsafe fn open(
        &self,
        features: Features,
        limits: &Limits,
    ) -> Result<OpenDevice<A>, DeviceError>;
    unsafe fn texture_format_capabilities(
        &self,
        format: TextureFormat,
    ) -> TextureFormatCapabilities;
    unsafe fn surface_capabilities(
        &self,
        surface: &<A as Api>::Surface,
    ) -> Option<SurfaceCapabilities>;
    unsafe fn get_presentation_timestamp(&self) -> PresentationTimestamp;
}

Required Methods

unsafe fn open( &self, features: Features, limits: &Limits, ) -> Result<OpenDevice<A>, DeviceError>

unsafe fn texture_format_capabilities( &self, format: TextureFormat, ) -> TextureFormatCapabilities

Return the set of supported capabilities for a texture format.

unsafe fn surface_capabilities( &self, surface: &<A as Api>::Surface, ) -> Option<SurfaceCapabilities>

Returns the capabilities of working with a specified surface.

None means presentation is not supported for it.

unsafe fn get_presentation_timestamp(&self) -> PresentationTimestamp

Creates a PresentationTimestamp using the adapter’s WSI.