Trait wgpu::hal::Surface

pub trait Surface<A>: WasmNotSendSync
where
    A: Api,
{
    // Required methods
    unsafe fn configure(
        &self,
        device: &<A as Api>::Device,
        config: &SurfaceConfiguration
    ) -> Result<(), SurfaceError>;
    unsafe fn unconfigure(&self, device: &<A as Api>::Device);
    unsafe fn acquire_texture(
        &self,
        timeout: Option<Duration>
    ) -> Result<Option<AcquiredSurfaceTexture<A>>, SurfaceError>;
    unsafe fn discard_texture(&self, texture: <A as Api>::SurfaceTexture);
}

Available on wgpu_core only.

Required Methods

unsafe fn configure( &self, device: &<A as Api>::Device, config: &SurfaceConfiguration ) -> Result<(), SurfaceError>

Configures the surface to use the given device.

Safety

All gpu work that uses the surface must have been completed.

All AcquiredSurfaceTextures must have been destroyed.

All Api::TextureViews derived from the AcquiredSurfaceTextures must have been destroyed.

All surfaces created using other devices must have been unconfigured before this call.

unsafe fn unconfigure(&self, device: &<A as Api>::Device)

Unconfigures the surface on the given device.

Safety

All gpu work that uses the surface must have been completed.

All AcquiredSurfaceTextures must have been destroyed.

All Api::TextureViews derived from the AcquiredSurfaceTextures must have been destroyed.

The surface must have been configured on the given device.

unsafe fn acquire_texture( &self, timeout: Option<Duration> ) -> Result<Option<AcquiredSurfaceTexture<A>>, SurfaceError>

Returns the next texture to be presented by the swapchain for drawing

A timeout of None means to wait indefinitely, with no timeout.

Portability

Some backends can’t support a timeout when acquiring a texture and the timeout will be ignored.

Returns None on timing out.

unsafe fn discard_texture(&self, texture: <A as Api>::SurfaceTexture)