wgpu

Module macros

Convenience macros

Macros

hal_type_dx12 🔒 Non-dx12

Helper macro to generate the documentation for dx12 hal methods, referencing the hal type.

hal_type_gles 🔒 gles

Helper macro to generate the documentation for gles hal methods, referencing the hal type.

hal_type_metal 🔒 Non-metal

Helper macro to generate the documentation for metal hal methods, referencing the hal type.

hal_type_vulkan 🔒 vulkan

Helper macro to generate the documentation for vulkan hal methods, referencing the hal type.