wgpu::naga

Module proc

Module processing functionality.

Modules

index

Definitions for index bounds checking.

Structs

Alignment

A newtype struct where its only valid values are powers of 2

BoundsCheckPolicies

Policies for injecting bounds checks during code generation.

ConstantEvaluator

A context for evaluating constant expressions.

Emitter

Helper class to emit expressions

ExpressionConstnessTracker

GlobalCtx

LayoutError

Layouter

Helper processor that derives the sizes of all types.

Namer

This processor assigns names to all the things in a module that may need identifiers in a textual backend.

ResolveContext

TypeLayout

Size and alignment information for a type.

Enums

BoundsCheckPolicy

How should code generated by Naga do bounds checks?

ConstantEvaluatorError

IndexableLength

The number of elements in an indexable type.

IndexableLengthError

LayoutErrorInner

NameKey

ResolveError

TypeResolution

The result of computing an expression’s type.

Constants

POINTER_SPAN

Functions

ensure_block_returns

Ensure that the given block has return statements at the end of its control flow.

flatten_compose

Return an iterator over the individual components assembled by a Compose expression.

Type Aliases

EntryPointIndex