Struct wgpu::BufferSlice

pub struct BufferSlice<'a> { /* fields omitted */ }

Slice into a Buffer.

Created by calling Buffer::slice. To use the whole buffer, call with unbounded slice:

buffer.slice(..)

Implementations

impl<'a> BufferSlice<'a>

pub fn map_async(
&self,
mode: MapMode
) -> impl Future<Output = Result<(), BufferAsyncError>> + Send

Map the buffer. Buffer is ready to map once the future is resolved.

For the future to complete, device.poll(...) must be called elsewhere in the runtime, possibly integrated into an event loop, run on a separate thread, or continually polled in the same task runtime that this future will be run on.

It’s expected that wgpu will eventually supply its own event loop infrastructure that will be easy to integrate into other event loops, like winit’s.

pub fn get_mapped_range(&self) -> BufferView<'a>

Synchronously and immediately map a buffer for reading. If the buffer is not immediately mappable through BufferDescriptor::mapped_at_creation or BufferSlice::map_async, will panic.

pub fn get_mapped_range_mut(&self) -> BufferViewMut<'a>

Synchronously and immediately map a buffer for writing. If the buffer is not immediately mappable through BufferDescriptor::mapped_at_creation or BufferSlice::map_async, will panic.