Struct wgpu::Device

pub struct Device { /* private fields */ }

Open connection to a graphics and/or compute device.

Responsible for the creation of most rendering and compute resources. These are then used in commands, which are submitted to a Queue.

A device may be requested from an adapter with Adapter::request_device.

Corresponds to WebGPU GPUDevice.

Implementations

impl Device

pub fn poll(&self, maintain: Maintain) -> bool

Check for resource cleanups and mapping callbacks.

Return true if the queue is empty, or false if there are more queue submissions still in flight. (Note that, unless access to the Queue is coordinated somehow, this information could be out of date by the time the caller receives it. Queues can be shared between threads, so other threads could submit new work at any time.)

On the web, this is a no-op. Devices are automatically polled.

pub fn features(&self) -> Features

List all features that may be used with this device.

Functions may panic if you use unsupported features.

pub fn limits(&self) -> Limits

List all limits that were requested of this device.

If any of these limits are exceeded, functions may panic.

pub fn create_shader_module(
&self,
desc: ShaderModuleDescriptor<'_>
) -> ShaderModule

Creates a shader module from either SPIR-V or WGSL source code.

pub unsafe fn create_shader_module_unchecked(
&self,
desc: ShaderModuleDescriptor<'_>
) -> ShaderModule

Creates a shader module from either SPIR-V or WGSL source code without runtime checks.

Safety

In contrast with create_shader_module this function creates a shader module without runtime checks which allows shaders to perform operations which can lead to undefined behavior like indexing out of bounds, thus it’s the caller responsibility to pass a shader which doesn’t perform any of this operations.

This has no effect on web.

pub unsafe fn create_shader_module_spirv(
&self,
desc: &ShaderModuleDescriptorSpirV<'_>
) -> ShaderModule

Creates a shader module from SPIR-V binary directly.

Safety

This function passes binary data to the backend as-is and can potentially result in a driver crash or bogus behaviour. No attempt is made to ensure that data is valid SPIR-V.

See also include_spirv_raw! and util::make_spirv_raw.

pub fn create_command_encoder(
&self,
desc: &CommandEncoderDescriptor<'_>
) -> CommandEncoder

Creates an empty CommandEncoder.

pub fn create_render_bundle_encoder(
&self,
desc: &RenderBundleEncoderDescriptor<'_>
) -> RenderBundleEncoder<'_>

Creates an empty RenderBundleEncoder.

pub fn create_bind_group(&self, desc: &BindGroupDescriptor<'_>) -> BindGroup

Creates a new BindGroup.

pub fn create_bind_group_layout(
&self,
desc: &BindGroupLayoutDescriptor<'_>
) -> BindGroupLayout

Creates a BindGroupLayout.

pub fn create_pipeline_layout(
&self,
desc: &PipelineLayoutDescriptor<'_>
) -> PipelineLayout

Creates a PipelineLayout.

pub fn create_render_pipeline(
&self,
desc: &RenderPipelineDescriptor<'_>
) -> RenderPipeline

Creates a RenderPipeline.

pub fn create_compute_pipeline(
&self,
desc: &ComputePipelineDescriptor<'_>
) -> ComputePipeline

Creates a ComputePipeline.

pub fn create_buffer(&self, desc: &BufferDescriptor<'_>) -> Buffer

Creates a Buffer.

pub fn create_texture(&self, desc: &TextureDescriptor<'_>) -> Texture

Creates a new Texture.

desc specifies the general format of the texture.

pub unsafe fn create_texture_from_hal<A: HalApi>(
&self,
hal_texture: A::Texture,
desc: &TextureDescriptor<'_>
) -> Texture

Creates a Texture from a wgpu-hal Texture.

Safety

hal_texture must be created from this device internal handle

hal_texture must be created respecting desc

hal_texture must be initialized

pub fn create_sampler(&self, desc: &SamplerDescriptor<'_>) -> Sampler

Creates a new Sampler.

desc specifies the behavior of the sampler.

pub fn create_query_set(&self, desc: &QuerySetDescriptor<'_>) -> QuerySet

Creates a new QuerySet.

pub fn on_uncaptured_error(&self, handler: impl UncapturedErrorHandler)

Set a callback for errors that are not handled in error scopes.

pub fn push_error_scope(&self, filter: ErrorFilter)

Push an error scope.

pub fn pop_error_scope(&self) -> impl Future<Output = Option<Error>> + Send

Pop an error scope.

pub fn start_capture(&self)

Starts frame capture.

pub fn stop_capture(&self)

Stops frame capture.

pub unsafe fn as_hal<A: HalApi, F: FnOnce(Option<&A::Device>) -> R, R>(
&self,
hal_device_callback: F
) -> R

Returns the inner hal Device using a callback. The hal device will be None if the backend type argument does not match with this wgpu Device

Safety

The raw handle obtained from the hal Device must not be manually destroyed