Struct wgpu::Features

#[repr(transparent)]
pub struct Features { /* fields omitted */ }

Features that are not guaranteed to be supported.

These are either part of the webgpu standard, or are extension features supported by wgpu when targeting native.

If you want to use a feature, you need to first verify that the adapter supports the feature. If the adapter does not support the feature, requesting a device with it enabled will panic.

Implementations

impl Features

pub const DEPTH_CLAMPING: Features

By default, polygon depth is clipped to 0-1 range. Anything outside of that range is rejected, and respective fragments are not touched.

With this extension, we can force clamping of the polygon depth to 0-1. That allows shadow map occluders to be rendered into a tighter depth range.

Supported platforms:

desktops

some mobile chips

This is a web and native feature.

pub const TEXTURE_COMPRESSION_BC: Features

Enables BCn family of compressed textures. All BCn textures use 4x4 pixel blocks with 8 or 16 bytes per block.

Compressed textures sacrifice some quality in exchange for significantly reduced bandwidth usage.

Support for this feature guarantees availability of [TextureUsage::COPY_SRC | TextureUsage::COPY_DST | TextureUsage::SAMPLED] for BCn formats. Features::TEXTURE_ADAPTER_SPECIFIC_FORMAT_FEATURES may enable additional usages.

Supported Platforms:

desktops

This is a web and native feature.

pub const TIMESTAMP_QUERY: Features

Enables use of Timestamp Queries. These queries tell the current gpu timestamp when all work before the query is finished. Call [CommandEncoder::write_timestamp], [RenderPassEncoder::write_timestamp], or [ComputePassEncoder::write_timestamp] to write out a timestamp.

They must be resolved using [CommandEncoder::resolve_query_sets] into a buffer, then the result must be multiplied by the timestamp period [Device::get_timestamp_period] to get the timestamp in nanoseconds. Multiple timestamps can then be diffed to get the time for operations between them to finish.

Due to gfx-hal limitations, this is only supported on vulkan for now.

Supported Platforms:

Vulkan (works)

DX12 (future)

This is a web and native feature.

pub const PIPELINE_STATISTICS_QUERY: Features

Enables use of Pipeline Statistics Queries. These queries tell the count of various operations performed between the start and stop call. Call [RenderPassEncoder::begin_pipeline_statistics_query] to start a query, then call [RenderPassEncoder::end_pipeline_statistics_query] to stop one.

They must be resolved using [CommandEncoder::resolve_query_sets] into a buffer. The rules on how these resolve into buffers are detailed in the documentation for PipelineStatisticsTypes.

Due to gfx-hal limitations, this is only supported on vulkan for now.

Supported Platforms:

Vulkan (works)

DX12 (future)

This is a web and native feature.

pub const MAPPABLE_PRIMARY_BUFFERS: Features

Webgpu only allows the MAP_READ and MAP_WRITE buffer usage to be matched with COPY_DST and COPY_SRC respectively. This removes this requirement.

This is only beneficial on systems that share memory between CPU and GPU. If enabled on a system that doesn’t, this can severely hinder performance. Only use if you understand the consequences.

Supported platforms:

All

This is a native only feature.

pub const TEXTURE_BINDING_ARRAY: Features

Allows the user to create uniform arrays of textures in shaders:

eg. uniform texture2D textures[10].

If Features::STORAGE_RESOURCE_BINDING_ARRAY is supported as well as this, the user may also create uniform arrays of storage textures.

eg. uniform image2D textures[10].

This capability allows them to exist and to be indexed by dynamically uniform values.

Supported platforms:

DX12

Metal (with MSL 2.0+ on macOS 10.13+)

Vulkan

This is a native only feature.

pub const BUFFER_BINDING_ARRAY: Features

Allows the user to create arrays of buffers in shaders:

eg. uniform myBuffer { .... } buffer_array[10].

This capability allows them to exist and to be indexed by dynamically uniform values.

If Features::STORAGE_RESOURCE_BINDING_ARRAY is supported as well as this, the user may also create arrays of storage buffers.

eg. buffer myBuffer { ... } buffer_array[10]

Supported platforms:

DX12

Vulkan

This is a native only feature.

pub const STORAGE_RESOURCE_BINDING_ARRAY: Features

Allows the user to create uniform arrays of storage buffers or textures in shaders, if resp. Features::BUFFER_BINDING_ARRAY or Features::TEXTURE_BINDING_ARRAY is supported.

This capability allows them to exist and to be indexed by dynamically uniform values.

Supported platforms:

Metal (with MSL 2.2+ on macOS 10.13+)

Vulkan

This is a native only feature.

pub const RESOURCE_BINDING_ARRAY_NON_UNIFORM_INDEXING: Features

Allows shaders to index resource arrays with dynamically non-uniform values:

eg. texture_array[vertex_data]

In order to use this capability, the corresponding GLSL extension must be enabled like so:

#extension GL_EXT_nonuniform_qualifier : require

and then used either as nonuniformEXT qualifier in variable declaration:

eg. layout(location = 0) nonuniformEXT flat in int vertex_data;

or as nonuniformEXT constructor:

eg. texture_array[nonuniformEXT(vertex_data)]

HLSL does not need any extension.

Supported platforms:

DX12

Metal (with MSL 2.0+ on macOS 10.13+)

Vulkan 1.2+ (or VK_EXT_descriptor_indexing)’s shaderSampledImageArrayNonUniformIndexing feature)

This is a native only feature.

pub const UNSIZED_BINDING_ARRAY: Features

Allows the user to create unsized uniform arrays of bindings:

eg. uniform texture2D textures[].

If this capability is supported, Features::RESOURCE_BINDING_ARRAY_NON_UNIFORM_INDEXING is very likely to also be supported

Supported platforms:

DX12

Vulkan 1.2+ (or VK_EXT_descriptor_indexing)’s runtimeDescriptorArray feature

This is a native only feature.

pub const MULTI_DRAW_INDIRECT: Features

Allows the user to call [RenderPass::multi_draw_indirect] and [RenderPass::multi_draw_indexed_indirect].

Allows multiple indirect calls to be dispatched from a single buffer.

Supported platforms:

DX12

Metal

Vulkan

This is a native only feature.

pub const MULTI_DRAW_INDIRECT_COUNT: Features

Allows the user to call [RenderPass::multi_draw_indirect_count] and [RenderPass::multi_draw_indexed_indirect_count].

This allows the use of a buffer containing the actual number of draw calls.

Supported platforms:

DX12

Vulkan 1.2+ (or VK_KHR_draw_indirect_count)

This is a native only feature.

pub const PUSH_CONSTANTS: Features

Allows the use of push constants: small, fast bits of memory that can be updated inside a [RenderPass].

Allows the user to call [RenderPass::set_push_constants], provide a non-empty array to [PipelineLayoutDescriptor], and provide a non-zero limit to Limits::max_push_constant_size.

A block of push constants can be declared with layout(push_constant) uniform Name {..} in shaders.

Supported platforms:

DX12

Vulkan

Metal

DX11 (emulated with uniforms)

OpenGL (emulated with uniforms)

This is a native only feature.

pub const ADDRESS_MODE_CLAMP_TO_BORDER: Features

Allows the use of AddressMode::ClampToBorder.

Supported platforms:

DX12

Vulkan

Metal (macOS 10.12+ only)

DX11

OpenGL

This is a web and native feature.

pub const NON_FILL_POLYGON_MODE: Features

Allows the user to set a non-fill polygon mode in PrimitiveState::polygon_mode

This allows drawing polygons/triangles as lines (wireframe) or points instead of filled

Supported platforms:

DX12

Vulkan

This is a native only feature.

pub const TEXTURE_COMPRESSION_ETC2: Features

Enables ETC family of compressed textures. All ETC textures use 4x4 pixel blocks. ETC2 RGB and RGBA1 are 8 bytes per block. RTC2 RGBA8 and EAC are 16 bytes per block.

Compressed textures sacrifice some quality in exchange for significantly reduced bandwidth usage.

Support for this feature guarantees availability of [TextureUsage::COPY_SRC | TextureUsage::COPY_DST | TextureUsage::SAMPLED] for ETC2 formats. Features::TEXTURE_ADAPTER_SPECIFIC_FORMAT_FEATURES may enable additional usages.

Supported Platforms:

Intel/Vulkan

Mobile (some)

This is a native-only feature.

pub const TEXTURE_COMPRESSION_ASTC_LDR: Features

Enables ASTC family of compressed textures. ASTC textures use pixel blocks varying from 4x4 to 12x12. Blocks are always 16 bytes.

Compressed textures sacrifice some quality in exchange for significantly reduced bandwidth usage.

Support for this feature guarantees availability of [TextureUsage::COPY_SRC | TextureUsage::COPY_DST | TextureUsage::SAMPLED] for ASTC formats. Features::TEXTURE_ADAPTER_SPECIFIC_FORMAT_FEATURES may enable additional usages.

Supported Platforms:

Intel/Vulkan

Mobile (some)

This is a native-only feature.

pub const TEXTURE_ADAPTER_SPECIFIC_FORMAT_FEATURES: Features

Enables device specific texture format features.

See TextureFormatFeatures for a listing of the features in question.

By default only texture format properties as defined by the WebGPU specification are allowed. Enabling this feature flag extends the features of each format to the ones supported by the current device. Note that without this flag, read/write storage access is not allowed at all.

This extension does not enable additional formats.

This is a native-only feature.

pub const SHADER_FLOAT64: Features

Enables 64-bit floating point types in SPIR-V shaders.

Note: even when supported by GPU hardware, 64-bit floating point operations are frequently between 16 and 64 times slower than equivalent operations on 32-bit floats.

Supported Platforms:

Vulkan

This is a native-only feature.

pub const VERTEX_ATTRIBUTE_64BIT: Features

Enables using 64-bit types for vertex attributes.

Requires SHADER_FLOAT64.

Supported Platforms: N/A

This is a native-only feature.

pub const CONSERVATIVE_RASTERIZATION: Features

Allows the user to set a overestimation-conservative-rasterization in PrimitiveState::conservative

Processing of degenerate triangles/lines is hardware specific. Only triangles are supported.

Supported platforms:

Vulkan

This is a native only feature.

pub const VERTEX_WRITABLE_STORAGE: Features

Enables bindings of writable storage buffers and textures visible to vertex shaders.

Note: some (tiled-based) platforms do not support vertex shaders with any side-effects.

Supported Platforms:

All

This is a native-only feature.

pub const CLEAR_COMMANDS: Features

Enables clear to zero for buffers & images.

Supported platforms:

All

This is a native only feature.

pub const SPIRV_SHADER_PASSTHROUGH: Features

Enables creating shader modules from SPIR-V binary data (unsafe).

SPIR-V data is not parsed or interpreted in any way; you can use [wgpu::make_spirv_raw!] to check for alignment and magic number when converting from raw bytes.

Supported platforms:

Vulkan, in case shader’s requested capabilities and extensions agree with Vulkan implementation.

This is a native only feature.

pub const ALL_WEBGPU: Features

Features which are part of the upstream WebGPU standard.

pub const ALL_NATIVE: Features

Features that are only available when targeting native (not web).

pub const fn empty() -> Features

Returns an empty set of flags

pub const fn all() -> Features

Returns the set containing all flags.

pub const fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u64) -> Option<Features>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u64) -> Features

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u64) -> Features

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Features) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Features) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: Features)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Features)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Features)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Features, value: bool)

Inserts or removes the specified flags depending on the passed value.