wgpu

Struct FeaturesWGPU

#[repr(transparent)]
pub struct FeaturesWGPU(<FeaturesWGPU as PublicFlags>::Internal);

Features that are not guaranteed to be supported.

Most of these are native-only extension features supported by wgpu only when targeting native. A few are intended to align with a proposed WebGPU extension, and one (EXTERNAL_TEXTURE) controls WebGPU-specified behavior that is not optional in the standard, but that we don’t want to make a crate::DownlevelFlags until the implementation is more complete. For all features see Features.

If you want to use a feature, you need to first verify that the adapter supports the feature. If the adapter does not support the feature, requesting a device with it enabled will panic.

Corresponds to WebGPU GPUFeatureName.

Tuple Fields
0: <FeaturesWGPU as PublicFlags>::Internal
Implementations

impl FeaturesWGPU

pub const SHADER_FLOAT32_ATOMIC: FeaturesWGPU

Allows shaders to use f32 atomic load, store, add, sub, and exchange.

Supported platforms:

Metal (with MSL 3.0+ and Apple7+/Mac2)

Vulkan (with VK_EXT_shader_atomic_float)

This is a native only feature.

pub const TEXTURE_FORMAT_16BIT_NORM: FeaturesWGPU

Enables normalized 16-bit texture formats.

Supported platforms:

Vulkan

DX12

Metal

This is a native only feature.

pub const TEXTURE_COMPRESSION_ASTC_HDR: FeaturesWGPU

Enables ASTC HDR family of compressed textures.

Compressed textures sacrifice some quality in exchange for significantly reduced bandwidth usage.

Support for this feature guarantees availability of [TextureUsages::COPY_SRC | TextureUsages::COPY_DST | TextureUsages::TEXTURE_BINDING] for ASTC formats with the HDR channel type. Features::TEXTURE_ADAPTER_SPECIFIC_FORMAT_FEATURES may enable additional usages.

Supported Platforms:

Metal

Vulkan

OpenGL

This is a native only feature.

pub const TEXTURE_ADAPTER_SPECIFIC_FORMAT_FEATURES: FeaturesWGPU

Enables device specific texture format features.

See TextureFormatFeatures for a listing of the features in question.

By default only texture format properties as defined by the WebGPU specification are allowed. Enabling this feature flag extends the features of each format to the ones supported by the current device. Note that without this flag, read/write storage access is not allowed at all.

This extension does not enable additional formats.

This is a native only feature.

pub const PIPELINE_STATISTICS_QUERY: FeaturesWGPU

Enables use of Pipeline Statistics Queries. These queries tell the count of various operations performed between the start and stop call. Call RenderPass::begin_pipeline_statistics_query to start a query, then call RenderPass::end_pipeline_statistics_query to stop one.

They must be resolved using CommandEncoder::resolve_query_set into a buffer. The rules on how these resolve into buffers are detailed in the documentation for PipelineStatisticsTypes.

Supported Platforms:

Vulkan

DX12

This is a native only feature with a proposal for the web.

pub const TIMESTAMP_QUERY_INSIDE_ENCODERS: FeaturesWGPU

Allows for timestamp queries directly on command encoders.

Implies Features::TIMESTAMP_QUERY is supported.

Additionally allows for timestamp writes on command encoders using CommandEncoder::write_timestamp.

Supported platforms:

Vulkan

DX12

Metal

This is a native only feature.

pub const TIMESTAMP_QUERY_INSIDE_PASSES: FeaturesWGPU

Allows for timestamp queries directly on command encoders.

Implies Features::TIMESTAMP_QUERY & Features::TIMESTAMP_QUERY_INSIDE_ENCODERS is supported.

Additionally allows for timestamp queries to be used inside render & compute passes using:

RenderPass::write_timestamp

ComputePass::write_timestamp

Supported platforms:

Vulkan

DX12

Metal (AMD & Intel, not Apple GPUs)

This is generally not available on tile-based rasterization GPUs.

This is a native only feature with a proposal for the web.

pub const MAPPABLE_PRIMARY_BUFFERS: FeaturesWGPU

Webgpu only allows the MAP_READ and MAP_WRITE buffer usage to be matched with COPY_DST and COPY_SRC respectively. This removes this requirement.

This is only beneficial on systems that share memory between CPU and GPU. If enabled on a system that doesn’t, this can severely hinder performance. Only use if you understand the consequences.

Supported platforms:

Vulkan

DX12

Metal

This is a native only feature.

pub const TEXTURE_BINDING_ARRAY: FeaturesWGPU

Allows the user to create uniform arrays of textures in shaders:

ex.

var textures: binding_array<texture_2d<f32>, 10> (WGSL)

uniform texture2D textures[10] (GLSL)

If Features::STORAGE_RESOURCE_BINDING_ARRAY is supported as well as this, the user may also create uniform arrays of storage textures.

ex.

var textures: array<texture_storage_2d<r32float, write>, 10> (WGSL)

uniform image2D textures[10] (GLSL)

This capability allows them to exist and to be indexed by dynamically uniform values.

Supported platforms:

DX12

Metal (with MSL 2.0+ on macOS 10.13+)

Vulkan

This is a native only feature.

pub const BUFFER_BINDING_ARRAY: FeaturesWGPU

Allows the user to create arrays of buffers in shaders:

ex.

var<uniform> buffer_array: array<MyBuffer, 10> (WGSL)

uniform myBuffer { ... } buffer_array[10] (GLSL)

This capability allows them to exist and to be indexed by dynamically uniform values.

If Features::STORAGE_RESOURCE_BINDING_ARRAY is supported as well as this, the user may also create arrays of storage buffers.

ex.

var<storage> buffer_array: array<MyBuffer, 10> (WGSL)

buffer myBuffer { ... } buffer_array[10] (GLSL)

Supported platforms:

Vulkan

This is a native only feature.

pub const STORAGE_RESOURCE_BINDING_ARRAY: FeaturesWGPU

Allows the user to create uniform arrays of storage buffers or textures in shaders, if resp. Features::BUFFER_BINDING_ARRAY or Features::TEXTURE_BINDING_ARRAY is supported.

This capability allows them to exist and to be indexed by dynamically uniform values.

Supported platforms:

Metal (with MSL 2.2+ on macOS 10.13+)

Vulkan

This is a native only feature.

pub const SAMPLED_TEXTURE_AND_STORAGE_BUFFER_ARRAY_NON_UNIFORM_INDEXING: FeaturesWGPU

Allows shaders to index sampled texture and storage buffer resource arrays with dynamically non-uniform values:

ex. texture_array[vertex_data]

In order to use this capability, the corresponding GLSL extension must be enabled like so:

#extension GL_EXT_nonuniform_qualifier : require

and then used either as nonuniformEXT qualifier in variable declaration:

ex. layout(location = 0) nonuniformEXT flat in int vertex_data;

or as nonuniformEXT constructor:

ex. texture_array[nonuniformEXT(vertex_data)]

WGSL and HLSL do not need any extension.

Supported platforms:

DX12

Metal (with MSL 2.0+ on macOS 10.13+)

Vulkan 1.2+ (or VK_EXT_descriptor_indexing)’s shaderSampledImageArrayNonUniformIndexing & shaderStorageBufferArrayNonUniformIndexing feature)

This is a native only feature.

pub const STORAGE_TEXTURE_ARRAY_NON_UNIFORM_INDEXING: FeaturesWGPU

Allows shaders to index storage texture resource arrays with dynamically non-uniform values:

ex. texture_array[vertex_data]

Supported platforms:

DX12

Metal (with MSL 2.0+ on macOS 10.13+)

Vulkan 1.2+ (or VK_EXT_descriptor_indexing)’s shaderStorageTextureArrayNonUniformIndexing feature)

This is a native only feature.

pub const PARTIALLY_BOUND_BINDING_ARRAY: FeaturesWGPU

Allows the user to create bind groups containing arrays with less bindings than the BindGroupLayout.

Supported platforms:

Vulkan

DX12

This is a native only feature.

pub const MULTI_DRAW_INDIRECT_COUNT: FeaturesWGPU

Allows the user to call RenderPass::multi_draw_indirect_count and RenderPass::multi_draw_indexed_indirect_count.

This allows the use of a buffer containing the actual number of draw calls. This feature being present also implies that all calls to RenderPass::multi_draw_indirect and RenderPass::multi_draw_indexed_indirect are not being emulated with a series of draw_indirect calls.

Supported platforms:

DX12

Vulkan 1.2+ (or VK_KHR_draw_indirect_count)

This is a native only feature.

pub const PUSH_CONSTANTS: FeaturesWGPU

Allows the use of push constants: small, fast bits of memory that can be updated inside a RenderPass.

Allows the user to call RenderPass::set_push_constants, provide a non-empty array to PipelineLayoutDescriptor, and provide a non-zero limit to Limits::max_push_constant_size.

A block of push constants can be declared in WGSL with var<push_constant>:

struct PushConstants { example: f32, }
var<push_constant> c: PushConstants;

In GLSL, this corresponds to layout(push_constant) uniform Name {..}.

Supported platforms:

DX12

Vulkan

Metal

OpenGL (emulated with uniforms)

This is a native only feature.

pub const ADDRESS_MODE_CLAMP_TO_ZERO: FeaturesWGPU

Allows the use of AddressMode::ClampToBorder with a border color of SamplerBorderColor::Zero.

Supported platforms:

DX12

Vulkan

Metal

OpenGL

This is a native only feature.

pub const ADDRESS_MODE_CLAMP_TO_BORDER: FeaturesWGPU

Allows the use of AddressMode::ClampToBorder with a border color other than SamplerBorderColor::Zero.

Supported platforms:

DX12

Vulkan

Metal (macOS 10.12+ only)

OpenGL

This is a native only feature.

pub const POLYGON_MODE_LINE: FeaturesWGPU

Allows the user to set PolygonMode::Line in PrimitiveState::polygon_mode

This allows drawing polygons/triangles as lines (wireframe) instead of filled

Supported platforms:

DX12

Vulkan

Metal

This is a native only feature.

pub const POLYGON_MODE_POINT: FeaturesWGPU

Allows the user to set PolygonMode::Point in PrimitiveState::polygon_mode

This allows only drawing the vertices of polygons/triangles instead of filled

Supported platforms:

Vulkan

This is a native only feature.

pub const CONSERVATIVE_RASTERIZATION: FeaturesWGPU

Allows the user to set a overestimation-conservative-rasterization in PrimitiveState::conservative

Processing of degenerate triangles/lines is hardware specific. Only triangles are supported.

Supported platforms:

Vulkan

This is a native only feature.

pub const VERTEX_WRITABLE_STORAGE: FeaturesWGPU

Enables bindings of writable storage buffers and textures visible to vertex shaders.

Note: some (tiled-based) platforms do not support vertex shaders with any side-effects.

Supported Platforms:

All

This is a native only feature.

pub const CLEAR_TEXTURE: FeaturesWGPU

Enables clear to zero for textures.

Supported platforms:

All

This is a native only feature.

pub const MULTIVIEW: FeaturesWGPU

Enables multiview render passes and builtin(view_index) in vertex shaders.

Supported platforms:

Vulkan

OpenGL (web only)

This is a native only feature.

pub const VERTEX_ATTRIBUTE_64BIT: FeaturesWGPU

Enables using 64-bit types for vertex attributes.

Requires SHADER_FLOAT64.

Supported Platforms: N/A

This is a native only feature.

pub const TEXTURE_ATOMIC: FeaturesWGPU

Enables image atomic fetch add, and, xor, or, min, and max for R32Uint and R32Sint textures.

Supported platforms:

Vulkan

DX12

Metal (with MSL 3.1+)

This is a native only feature.

pub const TEXTURE_FORMAT_NV12: FeaturesWGPU

Allows for creation of textures of format TextureFormat::NV12

Supported platforms:

DX12

Vulkan

This is a native only feature.

pub const TEXTURE_FORMAT_P010: FeaturesWGPU

Allows for creation of textures of format TextureFormat::P010

Supported platforms:

DX12

Vulkan

This is a native only feature.

pub const EXTERNAL_TEXTURE: FeaturesWGPU

Allows for the creation and usage of ExternalTextures, and bind group layouts containing external texture BindingTypes.

Conceptually this should really be a crate::DownlevelFlags as it corresponds to WebGPU’s GPUExternalTexture. However, the implementation is currently in-progress, and until it is complete we do not want applications to ignore adapters due to a missing downlevel flag, when they may not require this feature at all.

Supported platforms:

DX12

Metal

pub const EXPERIMENTAL_RAY_QUERY: FeaturesWGPU

THIS IS EXPERIMENTAL: Features enabled by this may have major bugs in it and are expected to be subject to breaking changes, suggestions for the API exposed by this should be posted on the ray-tracing issue

Allows for the creation of ray-tracing queries within shaders.

Supported platforms:

Vulkan

This is a native-only feature.

pub const SHADER_F64: FeaturesWGPU

Enables 64-bit floating point types in SPIR-V shaders.

Note: even when supported by GPU hardware, 64-bit floating point operations are frequently between 16 and 64 times slower than equivalent operations on 32-bit floats.

Supported Platforms:

Vulkan

This is a native only feature.

pub const SHADER_I16: FeaturesWGPU

Allows shaders to use i16. Not currently supported in naga, only available through spirv-passthrough.

Supported platforms:

Vulkan

This is a native only feature.

pub const SHADER_PRIMITIVE_INDEX: FeaturesWGPU

Enables builtin(primitive_index) in fragment shaders.

Note: enables geometry processing for pipelines using the builtin. This may come with a significant performance impact on some hardware. Other pipelines are not affected.

Supported platforms:

Vulkan

DX12

Metal (some)

OpenGL (some)

This is a native only feature.

pub const SHADER_EARLY_DEPTH_TEST: FeaturesWGPU

Allows shaders to use the early_depth_test attribute.

The attribute is applied to the fragment shader entry point. It can be used in two ways:

Force early depth/stencil tests:

@early_depth_test(force) (WGSL)

layout(early_fragment_tests) in; (GLSL)

Provide a conservative depth specifier that allows an additional early depth test under certain conditions:

@early_depth_test(greater_equal/less_equal/unchanged) (WGSL)

layout(depth_<greater/less/unchanged>) out float gl_FragDepth; (GLSL)

See EarlyDepthTest for more details.

Supported platforms:

Vulkan

GLES 3.1+

This is a native only feature.

pub const SHADER_INT64: FeaturesWGPU

Allows shaders to use i64 and u64.

Supported platforms:

Vulkan

DX12 (DXC only)

Metal (with MSL 2.3+)

This is a native only feature.

pub const SUBGROUP: FeaturesWGPU

Allows compute and fragment shaders to use the subgroup operation built-ins and perform subgroup operations (except barriers).

Supported Platforms:

Vulkan

DX12

Metal

This is a native only feature.

pub const SUBGROUP_VERTEX: FeaturesWGPU

Allows vertex shaders to use the subgroup operation built-ins and perform subgroup operations (except barriers).

Supported Platforms:

Vulkan

This is a native only feature.

pub const SUBGROUP_BARRIER: FeaturesWGPU

Allows compute shaders to use the subgroup barrier.

Requires Features::SUBGROUP. Without it, enables nothing.

Supported Platforms:

Vulkan

Metal

This is a native only feature.

pub const PIPELINE_CACHE: FeaturesWGPU

Allows the use of pipeline cache objects

Supported platforms:

Vulkan

Unimplemented Platforms:

DX12

Metal

pub const SHADER_INT64_ATOMIC_MIN_MAX: FeaturesWGPU

Allows shaders to use i64 and u64 atomic min and max.

Supported platforms:

Vulkan (with VK_KHR_shader_atomic_int64)

DX12 (with SM 6.6+)

Metal (with MSL 2.4+)

This is a native only feature.

pub const SHADER_INT64_ATOMIC_ALL_OPS: FeaturesWGPU

Allows shaders to use all i64 and u64 atomic operations.

Supported platforms:

Vulkan (with VK_KHR_shader_atomic_int64)

DX12 (with SM 6.6+)

This is a native only feature.

pub const VULKAN_GOOGLE_DISPLAY_TIMING: FeaturesWGPU

Allows using the VK_GOOGLE_display_timing Vulkan extension.

This is used for frame pacing to reduce latency, and is generally only available on Android.

This feature does not have a wgpu-level API, and so users of wgpu wishing to use this functionality must access it using various as_hal functions, primarily Surface::as_hal(), to then use.

Supported platforms:

Vulkan (with VK_GOOGLE_display_timing)

This is a native only feature.

pub const VULKAN_EXTERNAL_MEMORY_WIN32: FeaturesWGPU

Allows using the VK_KHR_external_memory_win32 Vulkan extension.

Supported platforms:

Vulkan (with VK_KHR_external_memory_win32)

This is a native only feature.

pub const TEXTURE_INT64_ATOMIC: FeaturesWGPU

Enables R64Uint image atomic min and max.

Supported platforms:

Vulkan (with VK_EXT_shader_image_atomic_int64)

DX12 (with SM 6.6+)

Metal (with MSL 3.1+)

This is a native only feature.

pub const UNIFORM_BUFFER_BINDING_ARRAYS: FeaturesWGPU

Allows uniform buffers to be bound as binding arrays.

This allows:

Shaders to contain var<uniform> buffer: binding_array<UniformBuffer>;

The count field of BindGroupLayoutEntrys with Uniform buffers, to be set to Some.

Supported platforms:

None (https://github.com/gfx-rs/wgpu/issues/7149)

Potential Platforms:

DX12

Metal

Vulkan 1.2+ (or VK_EXT_descriptor_indexing)’s shaderUniformBufferArrayNonUniformIndexing feature)

This is a native only feature.

pub const EXPERIMENTAL_MESH_SHADER: FeaturesWGPU

Enables mesh shaders and task shaders in mesh shader pipelines.

Supported platforms:

Vulkan (with VK_EXT_mesh_shader)

Potential Platforms:

DX12

Metal

This is a native only feature.

pub const EXPERIMENTAL_RAY_HIT_VERTEX_RETURN: FeaturesWGPU

THIS IS EXPERIMENTAL: Features enabled by this may have major bugs in them and are expected to be subject to breaking changes, suggestions for the API exposed by this should be posted on the ray-tracing issue

Allows for returning of the hit triangle’s vertex position when tracing with an acceleration structure marked with AccelerationStructureFlags::ALLOW_RAY_HIT_VERTEX_RETURN.

Supported platforms:

Vulkan

This is a native only feature

pub const EXPERIMENTAL_MESH_SHADER_MULTIVIEW: FeaturesWGPU

Enables multiview in mesh shader pipelines

Supported platforms:

Vulkan (with VK_EXT_mesh_shader)

Potential Platforms:

DX12

Metal

This is a native only feature.

pub const EXTENDED_ACCELERATION_STRUCTURE_VERTEX_FORMATS: FeaturesWGPU

Allows usage of additional vertex formats in BlasTriangleGeometrySizeDescriptor::vertex_format

Supported platforms

Vulkan

DX12

pub const EXPERIMENTAL_PASSTHROUGH_SHADERS: FeaturesWGPU

Enables creating shaders from passthrough with reflection info (unsafe)

Allows using Device::create_shader_module_passthrough. Shader code isn’t parsed or interpreted in any way. It is the user’s responsibility to ensure the code and reflection (if passed) are correct.

Supported platforms

Vulkan

DX12

Metal

WebGPU

Ideally, in the future, all platforms will be supported. For more info, see this comment.

impl FeaturesWGPU

pub const fn empty() -> FeaturesWGPU

Get a flags value with all bits unset.

pub const fn all() -> FeaturesWGPU

Get a flags value with all known bits set.

pub const fn bits(&self) -> u64

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u64) -> Option<FeaturesWGPU>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u64) -> FeaturesWGPU

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u64) -> FeaturesWGPU

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<FeaturesWGPU>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: FeaturesWGPU) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: FeaturesWGPU) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: FeaturesWGPU)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: FeaturesWGPU)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: FeaturesWGPU)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: FeaturesWGPU, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: FeaturesWGPU) -> FeaturesWGPU

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: FeaturesWGPU) -> FeaturesWGPU

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: FeaturesWGPU) -> FeaturesWGPU

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: FeaturesWGPU) -> FeaturesWGPU

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> FeaturesWGPU

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl FeaturesWGPU

pub const fn iter(&self) -> Iter<FeaturesWGPU>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<FeaturesWGPU>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.