wgpu

Struct FeaturesWebGPU

#[repr(transparent)]
pub struct FeaturesWebGPU(<FeaturesWebGPU as PublicFlags>::Internal);

Features that are not guaranteed to be supported.

These are part of the WebGPU standard. For all features, see Features.

If you want to use a feature, you need to first verify that the adapter supports the feature. If the adapter does not support the feature, requesting a device with it enabled will panic.

Corresponds to WebGPU GPUFeatureName.

Tuple Fields
0: <FeaturesWebGPU as PublicFlags>::Internal
Implementations

impl FeaturesWebGPU

pub const DEPTH_CLIP_CONTROL: FeaturesWebGPU

By default, polygon depth is clipped to 0-1 range before/during rasterization. Anything outside of that range is rejected, and respective fragments are not touched.

With this extension, we can disabling clipping. That allows shadow map occluders to be rendered into a tighter depth range.

Supported platforms:

desktops

some mobile chips

This is a web and native feature.

pub const DEPTH32FLOAT_STENCIL8: FeaturesWebGPU

Allows for explicit creation of textures of format TextureFormat::Depth32FloatStencil8

Supported platforms:

Vulkan (mostly)

DX12

Metal

OpenGL

This is a web and native feature.

pub const TEXTURE_COMPRESSION_BC: FeaturesWebGPU

Enables BCn family of compressed textures. All BCn textures use 4x4 pixel blocks with 8 or 16 bytes per block.

Compressed textures sacrifice some quality in exchange for significantly reduced bandwidth usage.

Support for this feature guarantees availability of [TextureUsages::COPY_SRC | TextureUsages::COPY_DST | TextureUsages::TEXTURE_BINDING] for BCn formats. Features::TEXTURE_ADAPTER_SPECIFIC_FORMAT_FEATURES may enable additional usages.

This feature guarantees availability of sliced-3d textures for BC formats when combined with TEXTURE_COMPRESSION_BC_SLICED_3D.

Supported Platforms:

desktops

Mobile (All Apple9 and some Apple7 and Apple8 devices)

This is a web and native feature.

pub const TEXTURE_COMPRESSION_BC_SLICED_3D: FeaturesWebGPU

Allows the 3d dimension for textures with BC compressed formats.

This feature must be used in combination with TEXTURE_COMPRESSION_BC to enable 3D textures with BC compression. It does not enable the BC formats by itself.

Supported Platforms:

desktops

Mobile (All Apple9 and some Apple7 and Apple8 devices)

This is a web and native feature.

pub const TEXTURE_COMPRESSION_ETC2: FeaturesWebGPU

Enables ETC family of compressed textures. All ETC textures use 4x4 pixel blocks. ETC2 RGB and RGBA1 are 8 bytes per block. RTC2 RGBA8 and EAC are 16 bytes per block.

Compressed textures sacrifice some quality in exchange for significantly reduced bandwidth usage.

Support for this feature guarantees availability of [TextureUsages::COPY_SRC | TextureUsages::COPY_DST | TextureUsages::TEXTURE_BINDING] for ETC2 formats. Features::TEXTURE_ADAPTER_SPECIFIC_FORMAT_FEATURES may enable additional usages.

Supported Platforms:

Vulkan on Intel

Mobile (some)

This is a web and native feature.

pub const TEXTURE_COMPRESSION_ASTC: FeaturesWebGPU

Enables ASTC family of compressed textures. ASTC textures use pixel blocks varying from 4x4 to 12x12. Blocks are always 16 bytes.

Compressed textures sacrifice some quality in exchange for significantly reduced bandwidth usage.

Support for this feature guarantees availability of [TextureUsages::COPY_SRC | TextureUsages::COPY_DST | TextureUsages::TEXTURE_BINDING] for ASTC formats with Unorm/UnormSrgb channel type. Features::TEXTURE_ADAPTER_SPECIFIC_FORMAT_FEATURES may enable additional usages.

This feature does not guarantee availability of sliced 3d textures for ASTC formats. If available, 3d support can be enabled by TEXTURE_COMPRESSION_ASTC_SLICED_3D feature.

Supported Platforms:

Vulkan on Intel

Mobile (some)

This is a web and native feature.

pub const TEXTURE_COMPRESSION_ASTC_SLICED_3D: FeaturesWebGPU

Allows the 3d dimension for textures with ASTC compressed formats.

This feature must be used in combination with TEXTURE_COMPRESSION_ASTC to enable 3D textures with ASTC compression. It does not enable the ASTC formats by itself.

Supported Platforms:

Vulkan (some)

Metal on Apple3+

OpenGL/WebGL (some)

Not Supported:

DX12

This is a web and native feature.

pub const TIMESTAMP_QUERY: FeaturesWebGPU

Enables use of Timestamp Queries. These queries tell the current gpu timestamp when all work before the query is finished.

This feature allows the use of

RenderPassDescriptor::timestamp_writes

ComputePassDescriptor::timestamp_writes to write out timestamps.

For arbitrary timestamp write commands on encoders refer to Features::TIMESTAMP_QUERY_INSIDE_ENCODERS. For arbitrary timestamp write commands on passes refer to Features::TIMESTAMP_QUERY_INSIDE_PASSES.

They must be resolved using CommandEncoder::resolve_query_set into a buffer, then the result must be multiplied by the timestamp period Queue::get_timestamp_period to get the timestamp in nanoseconds. Multiple timestamps can then be diffed to get the time for operations between them to finish.

Supported Platforms:

Vulkan

DX12

Metal

This is a web and native feature.

pub const INDIRECT_FIRST_INSTANCE: FeaturesWebGPU

Allows non-zero value for the first_instance member in indirect draw calls.

If this feature is not enabled, and the first_instance member is non-zero, the behavior may be:

The draw call is ignored.

The draw call is executed as if the first_instance is zero.

The draw call is executed with the correct first_instance value.

Supported Platforms:

Vulkan (mostly)

DX12

Metal on Apple3+ or Mac1+

OpenGL (Desktop 4.2+ with ARB_shader_draw_parameters only)

Not Supported:

OpenGL ES / WebGL

This is a web and native feature.

pub const SHADER_F16: FeaturesWebGPU

Allows shaders to use 16-bit floating point types. You may use them uniform buffers, storage buffers, and local variables. You may not use them in immediates.

In order to use this in WGSL shaders, you must add enable f16; to the top of your shader, before any global items.

Supported Platforms:

Vulkan

Metal

DX12

This is a web and native feature.

pub const RG11B10UFLOAT_RENDERABLE: FeaturesWebGPU

Allows for usage of textures of format TextureFormat::Rg11b10Ufloat as a render target

Supported platforms:

Vulkan

DX12

Metal

This is a web and native feature.

pub const BGRA8UNORM_STORAGE: FeaturesWebGPU

Allows the TextureUsages::STORAGE_BINDING usage on textures with format TextureFormat::Bgra8Unorm

Supported Platforms:

Vulkan

DX12

Metal

This is a web and native feature.

pub const FLOAT32_FILTERABLE: FeaturesWebGPU

Allows textures with formats “r32float”, “rg32float”, and “rgba32float” to be filterable.

Supported Platforms:

Vulkan (mainly on Desktop GPUs)

DX12

Metal on macOS or Apple9+ GPUs, optional on iOS/iPadOS with Apple7/8 GPUs

GL with one of GL_ARB_color_buffer_float/GL_EXT_color_buffer_float/OES_texture_float_linear

This is a web and native feature.

pub const DUAL_SOURCE_BLENDING: FeaturesWebGPU

Allows two outputs from a shader to be used for blending. Note that dual-source blending doesn’t support multiple render targets.

For more info see the OpenGL ES extension GL_EXT_blend_func_extended.

Supported platforms:

OpenGL ES (with GL_EXT_blend_func_extended)

Metal (with MSL 1.2+)

Vulkan (with dualSrcBlend)

DX12

This is a web and native feature.

pub const CLIP_DISTANCES: FeaturesWebGPU

Allows the use of @builtin(clip_distances) in WGSL.

Supported platforms:

Vulkan (mainly on Desktop GPUs)

GL (Desktop or GL_EXT_clip_cull_distance)

This is a web and native feature.

pub const IMMEDIATES: FeaturesWebGPU

Allows the use of immediate data: small, fast bits of memory that can be updated inside a RenderPass.

Allows the user to call RenderPass::set_immediates, provide a non-zero immediate data size to PipelineLayoutDescriptor, and provide a non-zero limit to Limits::max_immediate_size.

A block of immediate data can be declared in WGSL with var<immediate>:

struct Immediates { example: f32, }
var<immediate> c: Immediates;

In GLSL, this corresponds to layout(immediates) uniform Name {..}.

Supported platforms:

DX12

Vulkan

Metal

OpenGL (emulated with uniforms)

WebGPU support is currently a proposal and will be available in browsers in the future.

This is a web and native feature.

impl FeaturesWebGPU

pub const fn empty() -> FeaturesWebGPU

Get a flags value with all bits unset.

pub const fn all() -> FeaturesWebGPU

Get a flags value with all known bits set.

pub const fn bits(&self) -> u64

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u64) -> Option<FeaturesWebGPU>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u64) -> FeaturesWebGPU

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u64) -> FeaturesWebGPU

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<FeaturesWebGPU>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: FeaturesWebGPU) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: FeaturesWebGPU) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: FeaturesWebGPU)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: FeaturesWebGPU)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: FeaturesWebGPU)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: FeaturesWebGPU, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: FeaturesWebGPU) -> FeaturesWebGPU

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: FeaturesWebGPU) -> FeaturesWebGPU

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: FeaturesWebGPU) -> FeaturesWebGPU

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: FeaturesWebGPU) -> FeaturesWebGPU

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> FeaturesWebGPU

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl FeaturesWebGPU

pub const fn iter(&self) -> Iter<FeaturesWebGPU>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<FeaturesWebGPU>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.