Struct wgpu::ImageCopyTextureTaggedBase

pub struct ImageCopyTextureTaggedBase<T> {
    pub texture: T,
    pub mip_level: u32,
    pub origin: Origin3d,
    pub aspect: TextureAspect,
    pub color_space: PredefinedColorSpace,
    pub premultiplied_alpha: bool,
}

View of a texture which can be used to copy to a texture, including color space and alpha premultiplication information.

Corresponds to WebGPU GPUImageCopyTextureTagged.

Fields
texture: T

The texture to be copied to/from.

mip_level: u32

The target mip level of the texture.

origin: Origin3d

The base texel of the texture in the selected mip_level.

aspect: TextureAspect

The copy aspect.

color_space: PredefinedColorSpace

The color space of this texture.

premultiplied_alpha: bool

The premultiplication of this texture

Implementations

impl<T> ImageCopyTextureTagged<T>where T: Copy,

pub fn to_untagged(self) -> ImageCopyTexture<T>

Removes the colorspace information from the type.