Struct wgpu::Queue

pub struct Queue { /* fields omitted */ }

Handle to a command queue on a device.

A Queue executes recorded CommandBuffer objects and provides convenience methods for writing to buffers and textures.

Implementations

impl Queue

pub fn write_buffer(&self, buffer: &Buffer, offset: BufferAddress, data: &[u8])

Schedule a data write into buffer starting at offset.

pub fn write_texture(
&self,
texture: TextureCopyView<'_>,
data: &[u8],
data_layout: TextureDataLayout,
size: Extent3d
)

Schedule a data write into texture.

pub fn submit<I: IntoIterator<Item = CommandBuffer>>(&self, command_buffers: I)

Submits a series of finished command buffers for execution.