Struct wgpu::TextureViewDescriptorBase

#[repr(C)]pub struct TextureViewDescriptorBase<L> {
    pub label: L,
    pub format: TextureFormat,
    pub dimension: TextureViewDimension,
    pub aspect: TextureAspect,
    pub base_mip_level: u32,
    pub level_count: u32,
    pub base_array_layer: u32,
    pub array_layer_count: u32,
}

Describes a TextureView.

Fields
label: L

Debug label of the texture view. This will show up in graphics debuggers for easy identification.

format: TextureFormat

Format of the texture view. At this time, it must be the same as the underlying format of the texture.

dimension: TextureViewDimension

The dimension of the texture view. For 1D textures, this must be 1D. For 2D textures it must be one of D2, D2Array, Cube, and CubeArray. For 3D textures it must be 3D

aspect: TextureAspect

Aspect of the texture. Color textures must be TextureAspect::All.

base_mip_level: u32

Base mip level.

level_count: u32

Mip level count. Must be at least one. base_mip_level + level_count must be less or equal to underlying texture mip count.

base_array_layer: u32

Base array layer.

array_layer_count: u32

Layer count. Must be at least one. base_array_layer + array_layer_count must be less or equal to the underlying array count.

Implementations

impl<L> TextureViewDescriptor<L>

pub fn map_label<K>(
&self,
fun: impl FnOnce(&L) -> K
) -> TextureViewDescriptor<K>