wgpu

Struct Tlas

pub struct Tlas {
    pub(crate) inner: DispatchTlas,
    pub(crate) instances: Vec<Option<TlasInstance>>,
    pub(crate) lowest_unmodified: u32,
}

Top Level Acceleration Structure (TLAS).

A TLAS contains a series of TLAS instances, which are a reference to a BLAS and a transformation matrix placing the geometry in the world.

A TLAS also contains an extra set of TLAS instances in a device readable form, you cant interact directly with these, instead you have to build the TLAS with TLAS instances.

Fields
inner: DispatchTlasinstances: Vec<Option<TlasInstance>>lowest_unmodified: u32
Implementations

impl Tlas

pub unsafe fn as_hal<A: Api>( &mut self, ) -> Option<impl Deref<Target = A::AccelerationStructure>>

Available on wgpu_core only.

Get the wgpu_hal acceleration structure from this Tlas.

Find the Api struct corresponding to the active backend in wgpu_hal::api, and pass that struct to the to the A type parameter.

Returns a guard that dereferences to the type of the hal backend which implements A::AccelerationStructure.

Types

The returned type depends on the backend:

hal::api::Vulkan uses hal::vulkan::AccelerationStructure

hal::api::Metal uses hal::metal::AccelerationStructure

hal::api::Dx12 uses hal::dx12::AccelerationStructure

hal::api::Gles uses hal::gles::AccelerationStructure

Deadlocks

The returned guard holds a read-lock on a device-local “destruction” lock, which will cause all calls to destroy to block until the guard is released.

Errors

This method will return None if:

The acceleration structure is not from the backend specified by A.

The acceleration structure is from the webgpu or custom backend.

Safety

The returned resource must not be destroyed unless the guard is the last reference to it and it is not in use by the GPU. The guard and handle may be dropped at any time however.

All the safety requirements of wgpu-hal must be upheld.

pub fn as_custom<T: TlasInterface>(&self) -> Option<&T>

Available on custom only.

Returns custom implementation of Tlas (if custom backend and is internally T)

pub fn get(&self) -> &[Option<TlasInstance>]

Get a reference to all instances.

pub fn get_mut_slice( &mut self, range: Range<usize>, ) -> Option<&mut [Option<TlasInstance>]>

Get a mutable slice to a range of instances. Returns None if the range is out of bounds. All elements from the lowest accessed index up are marked as modified. For best performance it is recommended to prefer access to low elements and modify higher elements as little as possible. This can be done by ordering instances from the most to the least used. It is recommended to use Self::index_mut unless the option if out of bounds is required

pub fn get_mut_single( &mut self, index: usize, ) -> Option<&mut Option<TlasInstance>>

Get a single mutable reference to an instance. Returns None if the range is out of bounds. All elements from the lowest accessed index up are marked as modified. For best performance it is recommended to prefer access to low elements and modify higher elements as little as possible. This can be done by ordering instances from the most to the least used. It is recommended to use Self::index_mut unless the option if out of bounds is required

pub fn as_binding(&self) -> BindingResource<'_>

Get the binding resource for the underling acceleration structure, to be used when creating a BindGroup