wgpu

Trait WebGpuError

pub trait WebGpuError: Error + 'static {
    // Required method
    fn webgpu_error_type(&self) -> ErrorType;
}

A trait for querying the ErrorType classification of an error.

This is intended to be used as a convenience by implementations of WebGPU to classify errors returned by wgpu_core.

Required Methods

fn webgpu_error_type(&self) -> ErrorType

Determine the classification of this error as a WebGPU ErrorType.