wgpu

Type Alias ImageDataLayout

pub type ImageDataLayout = TexelCopyBufferLayout;

👎Deprecated since 24.0.0:
This has been renamed to TexelCopyBufferLayout, and will be removed in 25.0.0.

Old name for a TexelCopyBufferLayout.

Aliased Type

#[repr(C)]
pub struct ImageDataLayout {
    pub offset: u64,
    pub bytes_per_row: Option<u32>,
    pub rows_per_image: Option<u32>,
}

Fields
offset: u64

Offset into the buffer that is the start of the texture. Must be a multiple of texture block size. For non-compressed textures, this is 1.

bytes_per_row: Option<u32>

Bytes per “row” in an image.

A row is one row of pixels or of compressed blocks in the x direction.

This value is required if there are multiple rows (i.e. height or depth is more than one pixel or pixel block for compressed textures)

Must be a multiple of 256 for CommandEncoder::copy_buffer_to_texture and CommandEncoder::copy_texture_to_buffer. You must manually pad the image such that this is a multiple of 256. It will not affect the image data.

Queue::write_texture does not have this requirement.

Must be a multiple of the texture block size. For non-compressed textures, this is 1.

rows_per_image: Option<u32>

“Rows” that make up a single “image”.

A row is one row of pixels or of compressed blocks in the x direction.

An image is one layer in the z direction of a 3D image or 2DArray texture.

The amount of rows per image may be larger than the actual amount of rows of data.

Required if there are multiple images (i.e. the depth is more than one).