wgpu
Type Alias RequestAdapterOptions

pub type RequestAdapterOptions<'a, 'b> = RequestAdapterOptionsBase<&'a Surface<'b>>;

Additional information required when requesting an adapter.

For use with Instance::request_adapter.

Corresponds to WebGPU GPURequestAdapterOptions.

Aliased Type

struct RequestAdapterOptions<'a, 'b> {
    pub power_preference: PowerPreference,
    pub force_fallback_adapter: bool,
    pub compatible_surface: Option<&'a Surface<'b>>,
}

Fields
power_preference: PowerPreference

Power preference for the adapter.

force_fallback_adapter: bool

Indicates that only a fallback adapter can be returned. This is generally a “software” implementation on the system.

compatible_surface: Option<&'a Surface<'b>>

Surface that is required to be presentable with the requested adapter. This does not create the surface, only guarantees that the adapter can present to said surface. For WebGL, this is strictly required, as an adapter can not be created without a surface.