Module wgpu::util

Utility structures and functions.

Structs

BufferInitDescriptor

Describes a Buffer when allocating.

DispatchIndirect

The structure expected in indirect_buffer for ComputePass::dispatch_workgroups_indirect.

DownloadBuffer

CPU accessible buffer used to download data back from the GPU.

DrawIndexedIndirect

The structure expected in indirect_buffer for RenderEncoder::draw_indexed_indirect.

DrawIndirect

The structure expected in indirect_buffer for RenderEncoder::draw_indirect.

StagingBelt

Staging belt is a machine that uploads data.

Traits

DeviceExt

Utility methods not meant to be in the main API.

RenderEncoder

Methods shared by RenderPass and RenderBundleEncoder.

Functions

align_to

Aligns a value to an alignment.

backend_bits_from_env

Get a set of backend bits from the environment variable WGPU_BACKEND.

initialize_adapter_from_env

Initialize the adapter obeying the WGPU_ADAPTER_NAME environment variable.

initialize_adapter_from_env_or_default

Initialize the adapter obeying the WGPU_ADAPTER_NAME environment variable and if it doesn’t exist fall back on a default adapter.

make_spirv

Treat the given byte slice as a SPIR-V module.

make_spirv_raw

Version of make_spirv intended for use with Device::create_shader_module_spirv. Returns raw slice instead of ShaderSource.

parse_backends_from_comma_list

Generates a set of backends from a comma separated list of case-insensitive backend names.

power_preference_from_env

Get a power preference from the environment variable WGPU_POWER_PREF