Struct wgpu::util::StagingBelt

pub struct StagingBelt { /* fields omitted */ }

Staging belt is a machine that uploads data.

Internally it uses a ring-buffer of staging buffers that are sub-allocated. It has an advantage over Queue.write_buffer in a way that it returns a mutable slice, which you can fill to avoid an extra data copy.

Using a staging belt is slightly complicated, and generally goes as follows:

Write to buffers that need writing to using write_buffer.

Call finish.

Submit all command encoders used with write_buffer.

Call recall

Implementations

impl StagingBelt

pub fn new(chunk_size: BufferAddress) -> Self

Create a new staging belt.

The chunk_size is the unit of internal buffer allocation. It's better when it's big, but ideally still 1-4 times less than the total amount of data uploaded per submission.

pub fn write_buffer(
&mut self,
encoder: &mut CommandEncoder,
target: &Buffer,
offset: BufferAddress,
size: BufferSize,
device: &Device
) -> BufferViewMut<'_>

Allocate the staging belt slice of size to be uploaded into the target buffer at the specified offset.

The upload will be placed into the provided command encoder. This encoder must be submitted after finish is called and before recall is called.

pub fn finish(&mut self)

Prepare currently mapped buffers for use in a submission.

At this point, all the partially used staging buffers are closed until the GPU is done copying the data from them.

pub fn recall(&mut self) -> impl Future<Output = ()> + Send

Recall all of the closed buffers back to be reused.

This has to be called after the command encoders written to write_buffer are submitted!