Crate wgpu

A cross-platform graphics and compute library based on WebGPU.

Structs

Adapter
A handle to a physical graphics and/or compute device.

AdapterInfo
Metadata about a backend adapter.

BackendBit

BindGroup
An opaque handle to a binding group.

BindGroupDescriptor
A description of a group of bindings and the resources to be bound.

BindGroupLayout
An opaque handle to a binding group layout.

BindGroupLayoutBinding
A description of a single binding inside a bind group.

BindGroupLayoutDescriptor

Binding
A bindable resource and the slot to bind it to.

BlendDescriptor

Buffer
A handle to a GPU-accessible buffer.

BufferAsyncMapping

BufferCopyView
A view of a buffer which can be used to copy to or from a texture.

BufferDescriptor

BufferUsage

Color

ColorStateDescriptor

ColorWrite

CommandBuffer
An opaque handle to a command buffer on the GPU.

CommandBufferDescriptor

CommandEncoder
An object that encodes GPU operations.

CommandEncoderDescriptor

ComputePass
An in-progress recording of a compute pass.

ComputePipeline
A handle to a compute pipeline.

ComputePipelineDescriptor
A complete description of a compute pipeline.

CreateBufferMapped
A buffer being created, mapped in host memory.

DepthStencilStateDescriptor

Device
An open connection to a graphics and/or compute device.

DeviceDescriptor

Extensions

Extent3d

Limits

Origin3d

PipelineLayout
An opaque handle to a pipeline layout.

PipelineLayoutDescriptor
A description of a pipeline layout.

ProgrammableStageDescriptor
A description of a programmable pipeline stage.

Queue
A handle to a command queue on a device.

RasterizationStateDescriptor

RenderPass
An in-progress recording of a render pass.

RenderPassColorAttachmentDescriptor
A description of a color attachment.

RenderPassDepthStencilAttachmentDescriptor

RenderPassDescriptor
A description of all the attachments of a render pass.

RenderPipeline
A handle to a rendering (graphics) pipeline.

RenderPipelineDescriptor
A complete description of a render (graphics) pipeline.

RequestAdapterOptions

Sampler
A handle to a sampler.

SamplerDescriptor

ShaderModule
A handle to a compiled shader module.

ShaderModuleDescriptor

ShaderStage

StencilStateFaceDescriptor

Surface
A handle to a presentable surface.

SwapChain
A handle to a swap chain.

SwapChainDescriptor

SwapChainOutput
A swap chain image that can be rendered to.

Texture
A handle to a texture on the GPU.

TextureCopyView
A view of a texture which can be used to copy to or from a buffer or another texture.

TextureDescriptor

TextureUsage

TextureView
A handle to a texture view.

TextureViewDescriptor

VertexAttributeDescriptor

VertexBufferDescriptor
A description of a vertex buffer.

Enums

AddressMode

BindingResource
A resource that can be bound to a pipeline.

BindingType
Specific type of a binding..

BlendFactor

BlendOperation

BufferMapAsyncStatus

CompareFunction

CullMode

FilterMode

FrontFace

IndexFormat

InputStepMode

LoadOp

PowerPreference

PresentMode

PrimitiveTopology

StencilOperation

StoreOp

TextureAspect

TextureDimension

TextureFormat

TextureViewDimension

VertexFormat

Functions

read_spirv
Safely read SPIR-V

Type Definitions

BufferAddress

BufferMapAsyncResult

ShaderLocation