wgpu::api
Module buffer

Structs

Buffer

Handle to a GPU-accessible buffer.

BufferAsyncError

Error occurred when trying to async map a buffer.

BufferSlice

A slice of a Buffer, to be mapped, used for vertex or index data, or the like.

BufferView

A read-only view of a mapped buffer’s bytes.

BufferViewMut

A write-only view of a mapped buffer’s bytes.

MapContext 🔒

The mapped portion of a buffer, if any, and its outstanding views.

Enums

MapMode

Type of buffer mapping.

Functions

check_buffer_bounds 🔒

range_to_offset_size 🔒

Type Aliases

BufferDescriptor

Describes a Buffer.