wgpu::api

Module render_pass

Structs

Operations

Pair of load and store operations for an attachment aspect.

RenderPass

In-progress recording of a render pass: a list of render commands in a CommandEncoder.

RenderPassColorAttachment

Describes a color attachment to a RenderPass.

RenderPassDepthStencilAttachment

Describes a depth/stencil attachment to a RenderPass.

RenderPassDescriptor

Describes the attachments of a render pass.

RenderPassTimestampWrites

Describes the timestamp writes of a render pass.

Enums

LoadOp

Operation to perform to the output attachment at the start of a render pass.

StoreOp

Operation to perform to the output attachment at the end of a render pass.