wgpu

Module cmp

We need to impl PartialEq, Eq, PartialOrd, Ord, and Hash for all handle types in wgpu.

For types that have some already-unique property, we can use that property to implement these traits.

For types (like WebGPU) that don’t have such a property, we generate an identifier and use that.

Macros

impl_eq_ord_hash_arc_address 🔒

Implements PartialEq, Eq, PartialOrd, Ord, and Hash for a type by comparing the addresses of the Arcs.

impl_eq_ord_hash_proxy 🔒

Implements PartialEq, Eq, PartialOrd, Ord, and Hash for a type by proxying the operations to a single field.

Structs

AtomicU64supports_64bit_atomics

An integer type which can be safely shared between threads.

Identifier

Statics

NEXT_ID 🔒