Enum wgpu::BindingType

pub enum BindingType {
    Buffer {
        ty: BufferBindingType,
        has_dynamic_offset: bool,
        min_binding_size: Option<NonZeroU64>,
    },
    Sampler {
        filtering: bool,
        comparison: bool,
    },
    Texture {
        sample_type: TextureSampleType,
        view_dimension: TextureViewDimension,
        multisampled: bool,
    },
    StorageTexture {
        access: StorageTextureAccess,
        format: TextureFormat,
        view_dimension: TextureViewDimension,
    },
}

Specific type of a binding.

WebGPU spec: the enum of

https://gpuweb.github.io/gpuweb/#dictdef-gpubufferbindinglayout

https://gpuweb.github.io/gpuweb/#dictdef-gpusamplerbindinglayout

https://gpuweb.github.io/gpuweb/#dictdef-gputexturebindinglayout

https://gpuweb.github.io/gpuweb/#dictdef-gpustoragetexturebindinglayout

Variants

Buffer

A buffer binding.

Fields of Buffer

ty: BufferBindingType

Sub-type of the buffer binding.

has_dynamic_offset: bool

Indicates that the binding has a dynamic offset. One offset must be passed to [RenderPass::set_bind_group] for each dynamic binding in increasing order of binding number.

min_binding_size: Option<NonZeroU64>

Minimum size of the corresponding BufferBinding required to match this entry. When pipeline is created, the size has to cover at least the corresponding structure in the shader plus one element of the unbound array, which can only be last in the structure. If None, the check is performed at draw call time instead of pipeline and bind group creation.

Sampler

A sampler that can be used to sample a texture.

Example GLSL syntax:

layout(binding = 0)
uniform sampler s;

Fields of Sampler

filtering: bool

The sampling result is produced based on more than a single color sample from a texture, e.g. when bilinear interpolation is enabled.

A filtering sampler can only be used with a filterable texture.

comparison: bool

Use as a comparison sampler instead of a normal sampler. For more info take a look at the analogous functionality in OpenGL: https://www.khronos.org/opengl/wiki/Sampler_Object#Comparison_mode.

Texture

A texture binding.

Example GLSL syntax:

layout(binding = 0)
uniform texture2D t;

Fields of Texture

sample_type: TextureSampleType

Sample type of the texture binding.

view_dimension: TextureViewDimension

Dimension of the texture view that is going to be sampled.

multisampled: bool

True if the texture has a sample count greater than 1. If this is true, the texture must be read from shaders with texture1DMS, texture2DMS, or texture3DMS, depending on dimension.

StorageTexture

A storage texture.

Example GLSL syntax:

layout(set=0, binding=0, r32f) uniform image2D myStorageImage;

Note that the texture format must be specified in the shader as well. A list of valid formats can be found in the specification here: https://www.khronos.org/registry/OpenGL/specs/gl/GLSLangSpec.4.60.html#layout-qualifiers

Fields of StorageTexture

access: StorageTextureAccess

Allowed access to this texture.

format: TextureFormat

Format of the texture.

view_dimension: TextureViewDimension

Dimension of the texture view that is going to be sampled.

Implementations

impl BindingType

pub fn has_dynamic_offset(&self) -> bool

Returns true for buffer bindings with dynamic offset enabled.