Struct wgpu::Adapter

pub struct Adapter { /* fields omitted */ }

Handle to a physical graphics and/or compute device.

Adapters can be used to open a connection to the corresponding Device on the host system by using Adapter::request_device.

Does not have to be kept alive.

Implementations

impl Adapter

pub fn request_device(
&self,
desc: &DeviceDescriptor<'_>,
trace_path: Option<&Path>
) -> impl Future<Output = Result<(Device, Queue), RequestDeviceError>> + Send

Requests a connection to a physical device, creating a logical device.

Returns the Device together with a Queue that executes command buffers.

Arguments

desc - Description of the features and limits requested from the given device.

trace_path - Can be used for API call tracing, if that feature is enabled in wgpu-core.

Panics

Features specified by desc are not supported by this adapter.

Unsafe features were requested but not enabled when requesting the adapter.

Limits requested exceed the values provided by the adapter.

Adapter does not support all features wgpu requires to safely operate.

pub fn get_swap_chain_preferred_format(
&self,
surface: &Surface
) -> Option<TextureFormat>

Returns an optimal texture format to use for the SwapChain with this adapter.

Returns None if the surface is incompatible with the adapter.

pub fn features(&self) -> Features

List all features that are supported with this adapter.

Features must be explicitly requested in Adapter::request_device in order to use them.

pub fn limits(&self) -> Limits

List the “best” limits that are supported by this adapter.

Limits must be explicitly requested in Adapter::request_device to set the values that you are allowed to use.

pub fn get_info(&self) -> AdapterInfo

Get info about the adapter itself.

pub fn get_texture_format_features(
&self,
format: TextureFormat
) -> TextureFormatFeatures

Returns the features supported for a given texture format by this adapter.

Note that the WebGPU spec further restricts the available usages/features. To disable these restrictions on a device, request the Features::TEXTURE_ADAPTER_SPECIFIC_FORMAT_FEATURES feature.