Struct wgpu::CommandEncoder

pub struct CommandEncoder { /* private fields */ }

Encodes a series of GPU operations.

A command encoder can record RenderPasses, ComputePasses, and transfer operations between driver-managed resources like Buffers and Textures.

When finished recording, call CommandEncoder::finish to obtain a CommandBuffer which may be submitted for execution.

Corresponds to WebGPU GPUCommandEncoder.

Implementations

impl CommandEncoder

pub fn finish(self) -> CommandBuffer

Finishes recording and returns a CommandBuffer that can be submitted for execution.

pub fn begin_render_pass<'encoder>( &'encoder mut self, desc: &RenderPassDescriptor<'_> ) -> RenderPass<'encoder>

Begins recording of a render pass.

This function returns a RenderPass object which records a single render pass.

As long as the returned RenderPass has not ended, any mutating operation on this command encoder causes an error and invalidates it. Note that the 'encoder lifetime relationship protects against this, but it is possible to opt out of it by calling RenderPass::forget_lifetime. This can be useful for runtime handling of the encoder->pass dependency e.g. when pass and encoder are stored in the same data structure.

pub fn begin_compute_pass<'encoder>( &'encoder mut self, desc: &ComputePassDescriptor<'_> ) -> ComputePass<'encoder>

Begins recording of a compute pass.

This function returns a ComputePass object which records a single compute pass.

As long as the returned ComputePass has not ended, any mutating operation on this command encoder causes an error and invalidates it. Note that the 'encoder lifetime relationship protects against this, but it is possible to opt out of it by calling ComputePass::forget_lifetime. This can be useful for runtime handling of the encoder->pass dependency e.g. when pass and encoder are stored in the same data structure.

pub fn copy_buffer_to_buffer( &mut self, source: &Buffer, source_offset: BufferAddress, destination: &Buffer, destination_offset: BufferAddress, copy_size: BufferAddress )

Copy data from one buffer to another.

Panics

Buffer offsets or copy size not a multiple of COPY_BUFFER_ALIGNMENT.

Copy would overrun buffer.

Copy within the same buffer.

pub fn copy_buffer_to_texture( &mut self, source: TexelCopyBufferInfo<'_>, destination: TexelCopyTextureInfo<'_>, copy_size: Extent3d )

Copy data from a buffer to a texture.

pub fn copy_texture_to_buffer( &mut self, source: TexelCopyTextureInfo<'_>, destination: TexelCopyBufferInfo<'_>, copy_size: Extent3d )

Copy data from a texture to a buffer.

pub fn copy_texture_to_texture( &mut self, source: TexelCopyTextureInfo<'_>, destination: TexelCopyTextureInfo<'_>, copy_size: Extent3d )

Copy data from one texture to another.

Panics

Textures are not the same type

If a depth texture, or a multisampled texture, the entire texture must be copied

Copy would overrun either texture

pub fn clear_texture( &mut self, texture: &Texture, subresource_range: &ImageSubresourceRange )

Clears texture to zero.

Note that unlike with clear_buffer, COPY_DST usage is not required.

Implementation notes

implemented either via buffer copies and render/depth target clear, path depends on texture usages

behaves like texture zero init, but is performed immediately (clearing is not delayed via marking it as uninitialized)

Panics

CLEAR_TEXTURE extension not enabled

Range is out of bounds

pub fn clear_buffer( &mut self, buffer: &Buffer, offset: BufferAddress, size: Option<BufferAddress> )

Clears buffer to zero.

Panics

Buffer does not have COPY_DST usage.

Range is out of bounds

pub fn insert_debug_marker(&mut self, label: &str)

Inserts debug marker.

pub fn push_debug_group(&mut self, label: &str)

Start record commands and group it into debug marker group.

pub fn pop_debug_group(&mut self)

Stops command recording and creates debug group.

pub fn resolve_query_set( &mut self, query_set: &QuerySet, query_range: Range<u32>, destination: &Buffer, destination_offset: BufferAddress )

Resolves a query set, writing the results into the supplied destination buffer.

Occlusion and timestamp queries are 8 bytes each (see crate::QUERY_SIZE). For pipeline statistics queries, see PipelineStatisticsTypes for more information.

pub unsafe fn as_hal_mut<A: HalApi, F: FnOnce(Option<&mut A::CommandEncoder>) -> R, R>( &mut self, hal_command_encoder_callback: F ) -> Option<R>

Available on wgpu_core only.

Returns the inner hal CommandEncoder using a callback. The hal command encoder will be None if the backend type argument does not match with this wgpu CommandEncoder

This method will start the wgpu_core level command recording.

Safety

The raw handle obtained from the hal CommandEncoder must not be manually destroyed

impl CommandEncoder

Features::TIMESTAMP_QUERY_INSIDE_ENCODERS must be enabled on the device in order to call these functions.

pub fn write_timestamp(&mut self, query_set: &QuerySet, query_index: u32)

Issue a timestamp command at this point in the queue. The timestamp will be written to the specified query set, at the specified index.

Must be multiplied by Queue::get_timestamp_period to get the value in nanoseconds. Absolute values have no meaning, but timestamps can be subtracted to get the time it takes for a string of operations to complete.

Attention: Since commands within a command recorder may be reordered, there is no strict guarantee that timestamps are taken after all commands recorded so far and all before all commands recorded after. This may depend both on the backend and the driver.

impl CommandEncoder

Features::EXPERIMENTAL_RAY_TRACING_ACCELERATION_STRUCTURE must be enabled on the device in order to call these functions.

pub fn build_acceleration_structures<'a>( &mut self, blas: impl IntoIterator<Item = &'a BlasBuildEntry<'a>>, tlas: impl IntoIterator<Item = &'a TlasPackage> )

Build bottom and top level acceleration structures.

Builds the BLASes then the TLASes, but does not build the BLASes into the TLASes, that must be done by setting a TLAS instance in the TLAS package to one that contains the BLAS (and with an appropriate transform)

Validation

blas: Iterator of bottom level acceleration structure entries to build. For each entry, the provided size descriptor must be strictly smaller or equal to the descriptor given at BLAS creation, this means:

Less or equal number of geometries

Same kind of geometry (with index buffer or without) (same vertex/index format)

Same flags

Less or equal number of vertices

Less or equal number of indices (if applicable)

tlas: iterator of top level acceleration structure packages to build For each entry:

Each BLAS in each TLAS instance must have been being built in the current call or in a previous call to build_acceleration_structures or build_acceleration_structures_unsafe_tlas

The number of TLAS instances must be less than or equal to the max number of tlas instances when creating (if creating a package with TlasPackage::new() this is already satisfied)

If the device the command encoder is created from does not have Features::EXPERIMENTAL_RAY_TRACING_ACCELERATION_STRUCTURE enabled then a validation error is generated

A bottom level acceleration structure may be build and used as a reference in a top level acceleration structure in the same invocation of this function.

Bind group usage

When a top level acceleration structure is used in a bind group, some validation takes place:

The top level acceleration structure is valid and has been built.

All the bottom level acceleration structures referenced by the top level acceleration structure are valid and have been built prior, or at same time as the containing top level acceleration structure.

pub unsafe fn build_acceleration_structures_unsafe_tlas<'a>( &mut self, blas: impl IntoIterator<Item = &'a BlasBuildEntry<'a>>, tlas: impl IntoIterator<Item = &'a TlasBuildEntry<'a>> )

Build bottom and top level acceleration structures. See CommandEncoder::build_acceleration_structures for the safe version and more details. All validation in CommandEncoder::build_acceleration_structures except that listed under tlas applies here as well.

Safety

The contents of the raw instance buffer must be valid for the underling api.

All bottom level acceleration structures, referenced in the raw instance buffer must be valid and built, when the corresponding top level acceleration structure is built. (builds may happen in the same invocation of this function).

At the time when the top level acceleration structure is used in a bind group, all associated bottom level acceleration structures must be valid, and built (no later than the time when the top level acceleration structure was built).