Struct wgpu::ComputePass

pub struct ComputePass<'a> { /* fields omitted */ }

In-progress recording of a compute pass.

Implementations

impl<'a> ComputePass<'a>

pub fn set_bind_group(
&mut self,
index: u32,
bind_group: &'a BindGroup,
offsets: &[DynamicOffset]
)

Sets the active bind group for a given bind group index. The bind group layout in the active pipeline when the dispatch() function is called must match the layout of this bind group.

If the bind group have dynamic offsets, provide them in order of their declaration.

pub fn set_pipeline(&mut self, pipeline: &'a ComputePipeline)

Sets the active compute pipeline.

pub fn dispatch(&mut self, x: u32, y: u32, z: u32)

Dispatches compute work operations.

x, y and z denote the number of work groups to dispatch in each dimension.

pub fn dispatch_indirect(
&mut self,
indirect_buffer: &'a Buffer,
indirect_offset: BufferAddress
)

Dispatches compute work operations, based on the contents of the indirect_buffer.

impl<'a> ComputePass<'a>

Features::PUSH_CONSTANTS must be enabled on the device in order to call these functions.

pub fn set_push_constants(&mut self, offset: u32, data: &[u32])

Set push constant data.

Offset is measured in bytes, but must be a multiple of PUSH_CONSTANT_ALIGNMENT.

Data size must be a multiple of 4 and must be aligned to the 4s, so we take an array of u32. For example, with an offset of 4 and an array of [u32; 3], that will write to the range of 4..16.