Struct wgpu::ComputePass

pub struct ComputePass<'a> { /* fields omitted */ }

In-progress recording of a compute pass.

Implementations

impl<'a> ComputePass<'a>

pub fn set_bind_group(
&mut self,
index: u32,
bind_group: &'a BindGroup,
offsets: &[DynamicOffset]
)

Sets the active bind group for a given bind group index. The bind group layout in the active pipeline when the dispatch() function is called must match the layout of this bind group.

If the bind group have dynamic offsets, provide them in order of their declaration. These offsets have to be aligned to Limits::min_uniform_buffer_offset_alignment or Limits::min_storage_buffer_offset_alignment appropriately.

pub fn set_pipeline(&mut self, pipeline: &'a ComputePipeline)

Sets the active compute pipeline.

pub fn insert_debug_marker(&mut self, label: &str)

Inserts debug marker.

pub fn push_debug_group(&mut self, label: &str)

Start record commands and group it into debug marker group.

pub fn pop_debug_group(&mut self)

Stops command recording and creates debug group.

pub fn dispatch(&mut self, x: u32, y: u32, z: u32)

Dispatches compute work operations.

x, y and z denote the number of work groups to dispatch in each dimension.

pub fn dispatch_indirect(
&mut self,
indirect_buffer: &'a Buffer,
indirect_offset: BufferAddress
)

Dispatches compute work operations, based on the contents of the indirect_buffer.

The structure expected in indirect_buffer is the following:

// x, y and z denote the number of work groups to dispatch in each dimension.
#[repr(C)]
struct DispatchIndirect {
    x: u32,
    y: u32,
    z: u32,
}

impl<'a> ComputePass<'a>

Features::PUSH_CONSTANTS must be enabled on the device in order to call these functions.

pub fn set_push_constants(&mut self, offset: u32, data: &[u8])

Set push constant data.

Offset is measured in bytes, but must be a multiple of PUSH_CONSTANT_ALIGNMENT.

Data size must be a multiple of 4 and must be aligned to the 4s, so we take an array of u32. For example, with an offset of 4 and an array of [u32; 3], that will write to the range of 4..16.

impl<'a> ComputePass<'a>

Features::TIMESTAMP_QUERY must be enabled on the device in order to call these functions.

pub fn write_timestamp(&mut self, query_set: &QuerySet, query_index: u32)

Issue a timestamp command at this point in the queue. The timestamp will be written to the specified query set, at the specified index.

Must be multiplied by Queue::get_timestamp_period to get the value in nanoseconds. Absolute values have no meaning, but timestamps can be subtracted to get the time it takes for a string of operations to complete.

impl<'a> ComputePass<'a>

Features::PIPELINE_STATISTICS_QUERY must be enabled on the device in order to call these functions.

pub fn begin_pipeline_statistics_query(
&mut self,
query_set: &QuerySet,
query_index: u32
)

Start a pipeline statistics query on this render pass. It can be ended with end_pipeline_statistics_query. Pipeline statistics queries may not be nested.

pub fn end_pipeline_statistics_query(&mut self)

End the pipeline statistics query on this render pass. It can be started with begin_pipeline_statistics_query. Pipeline statistics queries may not be nested.