Struct wgpu::Device

pub struct Device { /* fields omitted */ }

Open connection to a graphics and/or compute device.

The Device is the responsible for the creation of most rendering and compute resources, as well as exposing Queue objects.

A device may be requested from an adapter with Adapter::request_device.

Implementations

impl Device

pub fn poll(&self, maintain: Maintain)

Check for resource cleanups and mapping callbacks.

no-op on the web, device is automatically polled.

pub fn features(&self) -> Features

List all features that may be used with this device.

Functions may panic if you use unsupported features.

pub fn limits(&self) -> Limits

List all limits that were requested of this device.

If any of these limits are exceeded, functions may panic.

pub fn create_shader_module(&self, source: ShaderModuleSource) -> ShaderModule

Creates a shader module from either SPIR-V or WGSL source code.

pub fn create_command_encoder(
&self,
desc: &CommandEncoderDescriptor
) -> CommandEncoder

Creates an empty CommandEncoder.

pub fn create_render_bundle_encoder(
&self,
desc: &RenderBundleEncoderDescriptor
) -> RenderBundleEncoder

Creates an empty RenderBundleEncoder.

pub fn create_bind_group(&self, desc: &BindGroupDescriptor) -> BindGroup

Creates a new BindGroup.

pub fn create_bind_group_layout(
&self,
desc: &BindGroupLayoutDescriptor
) -> BindGroupLayout

Creates a BindGroupLayout.

pub fn create_pipeline_layout(
&self,
desc: &PipelineLayoutDescriptor
) -> PipelineLayout

Creates a PipelineLayout.

pub fn create_render_pipeline(
&self,
desc: &RenderPipelineDescriptor
) -> RenderPipeline

Creates a RenderPipeline.

pub fn create_compute_pipeline(
&self,
desc: &ComputePipelineDescriptor
) -> ComputePipeline

Creates a ComputePipeline.

pub fn create_buffer(&self, desc: &BufferDescriptor) -> Buffer

Creates a Buffer.

pub fn create_buffer_with_data(&self, data: &[u8], usage: BufferUsage) -> Buffer

Creates a new buffer, maps it into host-visible memory, copies data from the given slice, and finally unmaps it, returning a Buffer.

pub fn create_texture(&self, desc: &TextureDescriptor) -> Texture

Creates a new Texture.

desc specifies the general format of the texture.

pub fn create_sampler(&self, desc: &SamplerDescriptor) -> Sampler

Creates a new Sampler.

desc specifies the behavior of the sampler.

pub fn create_swap_chain(
&self,
surface: &Surface,
desc: &SwapChainDescriptor
) -> SwapChain

Create a new SwapChain which targets surface.

Panics

A old SwapChainFrame is still alive referencing an old swapchain.

Texture format requested is unsupported on the swap chain.