Struct wgpu::Queue

pub struct Queue { /* fields omitted */ }

Handle to a command queue on a device.

A Queue executes recorded CommandBuffer objects and provides convenience methods for writing to buffers and textures.

Implementations

impl Queue

pub fn write_buffer(&self, buffer: &Buffer, offset: BufferAddress, data: &[u8])

Schedule a data write into buffer starting at offset.

This method is intended to have low performance costs. As such, the write is not immediately submitted, and instead enqueued internally to happen at the start of the next submit() call.

pub fn write_texture(
&self,
texture: ImageCopyTexture<'_>,
data: &[u8],
data_layout: ImageDataLayout,
size: Extent3d
)

Schedule a data write into texture.

This method is intended to have low performance costs. As such, the write is not immediately submitted, and instead enqueued internally to happen at the start of the next submit() call.

pub fn submit<I: IntoIterator<Item = CommandBuffer>>(&self, command_buffers: I)

Submits a series of finished command buffers for execution.

pub fn get_timestamp_period(&self) -> f32

Gets the amount of nanoseconds each tick of a timestamp query represents.

Returns zero if timestamp queries are unsupported.