Struct wgpu::Queue

pub struct Queue { /* private fields */ }

Handle to a command queue on a device.

A Queue executes recorded CommandBuffer objects and provides convenience methods for writing to buffers and textures. It can be created along with a Device by calling Adapter::request_device.

Corresponds to WebGPU GPUQueue.

Implementations

impl Queue

pub fn write_buffer(&self, buffer: &Buffer, offset: BufferAddress, data: &[u8])

Schedule a data write into buffer starting at offset.

This method is intended to have low performance costs. As such, the write is not immediately submitted, and instead enqueued internally to happen at the start of the next submit() call.

This method fails if data overruns the size of buffer starting at offset.

pub fn write_buffer_with<'a>( &'a self, buffer: &'a Buffer, offset: BufferAddress, size: BufferSize ) -> Option<QueueWriteBufferView<'a>>

Schedule a data write into buffer starting at offset via the returned QueueWriteBufferView.

Reading from this buffer is slow and will not yield the actual contents of the buffer.

This method is intended to have low performance costs. As such, the write is not immediately submitted, and instead enqueued internally to happen at the start of the next submit() call.

This method fails if size is greater than the size of buffer starting at offset.

pub fn write_texture( &self, texture: ImageCopyTexture<'_>, data: &[u8], data_layout: ImageDataLayout, size: Extent3d )

Schedule a write of some data into a texture.

data contains the texels to be written, which must be in the same format as the texture.

data_layout describes the memory layout of data, which does not necessarily have to have tightly packed rows.

texture specifies the texture to write into, and the location within the texture (coordinate offset, mip level) that will be overwritten.

size is the size, in texels, of the region to be written.

This method is intended to have low performance costs. As such, the write is not immediately submitted, and instead enqueued internally to happen at the start of the next submit() call. However, data will be immediately copied into staging memory; so the caller may discard it any time after this call completes.

This method fails if size overruns the size of texture, or if data is too short.

pub fn submit<I: IntoIterator<Item = CommandBuffer>>( &self, command_buffers: I ) -> SubmissionIndex

Submits a series of finished command buffers for execution.

pub fn get_timestamp_period(&self) -> f32

Gets the amount of nanoseconds each tick of a timestamp query represents.

Returns zero if timestamp queries are unsupported.

pub fn on_submitted_work_done(&self, callback: impl FnOnce() + Send + 'static)

Registers a callback when the previous call to submit finishes running on the gpu. This callback being called implies that all mapped buffer callbacks attached to the same submission have also been called.

For the callback to complete, either queue.submit(..), instance.poll_all(..), or device.poll(..) must be called elsewhere in the runtime, possibly integrated into an event loop or run on a separate thread.

The callback will be called on the thread that first calls the above functions after the gpu work has completed. There are no restrictions on the code you can run in the callback, however on native the call to the function will not complete until the callback returns, so prefer keeping callbacks short and used to set flags, send messages, etc.