wgpu
Type Alias PollType

pub type PollType = PollType<SubmissionIndex>;

Passed to Device::poll to control how and if it should block.

Aliased Type

enum PollType {
    WaitForSubmissionIndex(SubmissionIndex),
    Wait,
    Poll,
}

Variants

WaitForSubmissionIndex(SubmissionIndex)

On wgpu-core based backends, block until the given submission has completed execution, and any callbacks have been invoked.

On WebGPU, this has no effect. Callbacks are invoked from the window event loop.

Wait

Same as WaitForSubmissionIndex but waits for the most recent submission.

Poll

Check the device for a single time without blocking.