wgpu

Type Alias SamplerDescriptor

pub type SamplerDescriptor<'a> = SamplerDescriptor<Label<'a>>;

Describes a Sampler.

For use with Device::create_sampler.

Corresponds to WebGPU GPUSamplerDescriptor.

Aliased Type

pub struct SamplerDescriptor<'a> {
    pub label: Option<&'a str>,
    pub address_mode_u: AddressMode,
    pub address_mode_v: AddressMode,
    pub address_mode_w: AddressMode,
    pub mag_filter: FilterMode,
    pub min_filter: FilterMode,
    pub mipmap_filter: MipmapFilterMode,
    pub lod_min_clamp: f32,
    pub lod_max_clamp: f32,
    pub compare: Option<CompareFunction>,
    pub anisotropy_clamp: u16,
    pub border_color: Option<SamplerBorderColor>,
}

Fields
label: Option<&'a str>

Debug label of the sampler. This will show up in graphics debuggers for easy identification.

address_mode_u: AddressMode

How to deal with out of bounds accesses in the u (i.e. x) direction

address_mode_v: AddressMode

How to deal with out of bounds accesses in the v (i.e. y) direction

address_mode_w: AddressMode

How to deal with out of bounds accesses in the w (i.e. z) direction

mag_filter: FilterMode

How to filter the texture when it needs to be magnified (made larger)

min_filter: FilterMode

How to filter the texture when it needs to be minified (made smaller)

mipmap_filter: MipmapFilterMode

How to filter between mip map levels

lod_min_clamp: f32

Minimum level of detail (i.e. mip level) to use

lod_max_clamp: f32

Maximum level of detail (i.e. mip level) to use

compare: Option<CompareFunction>

If this is enabled, this is a comparison sampler using the given comparison function.

anisotropy_clamp: u16

Must be at least 1. If this is not 1, all filter modes must be linear.

border_color: Option<SamplerBorderColor>

Border color to use when address_mode is [AddressMode::ClampToBorder]