wgpu::util::init

Function is_browser_webgpu_supported

pub async fn is_browser_webgpu_supported() -> bool

Determines whether the Backends::BROWSER_WEBGPU backend is supported.

The result can only be true if this is called from the main thread or a dedicated worker. For convenience, this is also supported on non-wasm targets, always returning false there.