wgpu::util

Struct StagingBelt

pub struct StagingBelt {
    device: Device,
    chunk_size: BufferAddress,
    buffer_usages: BufferUsages,
    active_chunks: Vec<Chunk>,
    closed_chunks: Vec<Chunk>,
    free_chunks: Vec<Chunk>,
    sender: Exclusive<Sender<Chunk>>,
    receiver: Exclusive<Receiver<Chunk>>,
}

Available on std only.

Efficiently performs many buffer writes by sharing and reusing temporary buffers.

Internally it uses a ring-buffer of staging buffers that are sub-allocated. Its advantage over Queue::write_buffer_with() is that the individual allocations are cheaper; StagingBelt is most useful when you are writing very many small pieces of data. It can be understood as a sort of arena allocator.

Using a staging belt is slightly complicated, and generally goes as follows:

Use StagingBelt::write_buffer() or StagingBelt::allocate() to allocate buffer slices, then write your data to them.

Call StagingBelt::finish().

Submit all command encoders that were used in step 1.

Call StagingBelt::recall().

Fields
device: Devicechunk_size: BufferAddressbuffer_usages: BufferUsages

User-specified BufferUsages used to create the chunk buffers are created.

new guarantees that this always contains MAP_WRITE.

active_chunks: Vec<Chunk>

Chunks into which we are accumulating data to be transferred.

closed_chunks: Vec<Chunk>

Chunks that have scheduled transfers already; they are unmapped and some command encoder has one or more commands with them as source.

free_chunks: Vec<Chunk>

Chunks that are back from the GPU and ready to be mapped for write and put into active_chunks.

sender: Exclusive<Sender<Chunk>>

When closed chunks are mapped again, the map callback sends them here.

receiver: Exclusive<Receiver<Chunk>>

Free chunks are received here to be put on self.free_chunks.

Implementations

impl StagingBelt

pub fn new(device: Device, chunk_size: BufferAddress) -> Self

Create a new staging belt.

The chunk_size is the unit of internal buffer allocation; writes will be sub-allocated within each chunk. Therefore, for optimal use of memory, the chunk size should be:

larger than the largest single StagingBelt::write_buffer() operation;

1-4 times less than the total amount of data uploaded per submission (per StagingBelt::finish()); and

bigger is better, within these bounds.

The buffers returned by this StagingBelt will be have the buffer usages COPY_SRC | MAP_WRITE

pub fn new_with_buffer_usages( device: Device, chunk_size: BufferAddress, buffer_usages: BufferUsages, ) -> Self

Create a new staging belt.

The chunk_size is the unit of internal buffer allocation; writes will be sub-allocated within each chunk. Therefore, for optimal use of memory, the chunk size should be:

larger than the largest single StagingBelt::write_buffer() operation;

1-4 times less than the total amount of data uploaded per submission (per StagingBelt::finish()); and

bigger is better, within these bounds.

buffer_usages specifies the BufferUsages the staging buffers will be created with. MAP_WRITE will be added automatically. The method will panic if the combination of usages is not supported. Because MAP_WRITE is implied, the allowed usages depends on if Features::MAPPABLE_PRIMARY_BUFFERS is enabled.

If enabled: any usage is valid.

If disabled: only COPY_SRC can be used.

pub fn write_buffer( &mut self, encoder: &mut CommandEncoder, target: &Buffer, offset: BufferAddress, size: BufferSize, ) -> BufferViewMut

Allocate a staging belt slice of size to be copied into the target buffer at the specified offset.

offset and size must be multiples of COPY_BUFFER_ALIGNMENT (as is required by the underlying buffer operations).

The upload will be placed into the provided command encoder. This encoder must be submitted after StagingBelt::finish() is called and before StagingBelt::recall() is called.

If the size is greater than the size of any free internal buffer, a new buffer will be allocated for it. Therefore, the chunk_size passed to StagingBelt::new() should ideally be larger than every such size.

pub fn allocate( &mut self, size: BufferSize, alignment: BufferSize, ) -> BufferSlice<'_>

Allocate a staging belt slice with the given size and alignment and return it.

size must be a multiple of COPY_BUFFER_ALIGNMENT (as is required by the underlying buffer operations).

To use this slice, call BufferSlice::get_mapped_range_mut() and write your data into that BufferViewMut. (The view must be dropped before StagingBelt::finish() is called.)

You can then record your own GPU commands to perform with the slice, such as copying it to a texture (whereas StagingBelt::write_buffer() can only write to other buffers). All commands involving this slice must be submitted after StagingBelt::finish() is called and before StagingBelt::recall() is called.

If the size is greater than the space available in any free internal buffer, a new buffer will be allocated for it. Therefore, the chunk_size passed to StagingBelt::new() should ideally be larger than every such size.

The chosen slice will be positioned within the buffer at a multiple of alignment, which may be used to meet alignment requirements for the operation you wish to perform with the slice. This does not necessarily affect the alignment of the BufferViewMut.

pub fn finish(&mut self)

Prepare currently mapped buffers for use in a submission.

This must be called before the command encoder(s) provided to StagingBelt::write_buffer() are submitted.

At this point, all the partially used staging buffers are closed (cannot be used for further writes) until after StagingBelt::recall() is called and the GPU is done copying the data from them.

pub fn recall(&mut self)

Recall all of the closed buffers back to be reused.

This must only be called after the command encoder(s) provided to StagingBelt::write_buffer() are submitted. Additional calls are harmless. Not calling this as soon as possible may result in increased buffer memory usage.

fn receive_chunks(&mut self)

Move all chunks that the GPU is done with (and are now mapped again) from self.receiver to self.free_chunks.